% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{along_sum}
\alias{along_sum}
\title{Along-matrix summing function}
\usage{
along_sum(rpos, iterations)
}
\arguments{
\item{rpos}{Position vector}

\item{iterations}{Number of iterations}
}
\value{
A numeric matrix containing the adjacent column sums of
\code{rpos}.
}
\description{
Along-matrix summing function
}
\examples{
positions <- matrix(1:8, nrow = 2)
along_sum(positions, ncol(positions))
}
