% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{anderson_initialize}
\alias{anderson_initialize}
\title{Initialize GAS-object for modal series solution.}
\usage{
anderson_initialize(
  object,
  frequency,
  radius_body = NULL,
  g_body = NULL,
  h_body = NULL,
  sound_speed_sw = 1500,
  density_sw = 1026,
  ka_limit = NULL
)
}
\arguments{
\item{object}{GAS-class object.}

\item{frequency}{Frequency (Hz).}

\item{radius_body}{Radius of sphere (m).}

\item{g_body}{Density contrast for gas.}

\item{h_body}{Sound speed contrast for gas.}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}

\item{ka_limit}{Modal series limit (i.e. max "m"). The default is the maximum
ka + 10.}
}
\description{
Initialize GAS-object for modal series solution.
}
