% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{Arbitrary}
\alias{Arbitrary}
\alias{Arbitrary-class}
\title{Arbitrary body shape}
\value{
An object of the \code{Arbitrary} S4 class.
}
\description{
Arbitrary body shape
}
\examples{
methods::getClass("Arbitrary")

}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
