% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{available_models}
\alias{available_models}
\title{List available target-strength models}
\usage{
available_models()
}
\value{
A data frame describing currently available built-in and user-registered
target-strength models.
}
\description{
List available target-strength models
}
\examples{
head(available_models())
}
