% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{BBF}
\alias{BBF}
\alias{BBF-class}
\title{Backboned fish (BBF) object/class.}
\value{
An object of the \code{BBF} S4 class.
}
\description{
A S4 class that stores a fluid-like flesh component together with an explicit
elastic backbone component for swimbladder-less fish applications where those
two structures should remain separate. The body and backbone each retain
their own geometry, orientations, and material properties. This class is
intended to support composite flesh-plus-backbone model development and
future extensions where additional internal structures may need to be carried
alongside the body.
}
\examples{
methods::getClass("BBF")

}
\seealso{
\link{Scatterer}, \link{CSC}
}
\keyword{scatterer_type}
