% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bladder_rotation}
\alias{bladder_rotation}
\title{Support rotation function for KRM (swimbladder)}
\usage{
bladder_rotation(sum_rpos, rpos, theta, k_length)
}
\arguments{
\item{sum_rpos}{Summed position matrix}

\item{rpos}{Position matrix}

\item{theta}{Orientation angle}

\item{k_length}{Length of wavenumber vector}
}
\description{
Support rotation function for KRM (swimbladder)
}
