% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{body_rotation}
\alias{body_rotation}
\title{Support rotating function for KRM (body)}
\usage{
body_rotation(sum_rpos, rpos, theta, k_length)
}
\arguments{
\item{sum_rpos}{Summed position matrix}

\item{rpos}{Position matrix}

\item{theta}{Orientation angle}

\item{k_length}{Length of wavenumber vector}
}
\description{
Support rotating function for KRM (body)
}
