% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{brake_df}
\alias{brake_df}
\title{Support function for bending scatterer position matrix dataframe}
\usage{
brake_df(body_df, radius_curvature, mode = "ratio")
}
\arguments{
\item{body_df}{Dataframe object containing body shape information}

\item{radius_curvature}{Radius of curvature that can be parameterized either
as a ratio relative to body length or actual measurement}

\item{mode}{Either "ratio" or "measurement"}
}
\description{
Support function for bending scatterer position matrix dataframe
}
