% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{brake_scatterer}
\alias{brake_scatterer}
\title{Support function for bending scatterer body shape scatterer object}
\usage{
brake_scatterer(object, radius_curvature, mode = "ratio")
}
\arguments{
\item{object}{Scatterer-class object}

\item{radius_curvature}{Radius of curvature that can be parameterized either
as a ratio relative to body length or actual measurement}

\item{mode}{Either "ratio" or "measurement"}
}
\description{
Support function for bending scatterer body shape scatterer object
}
