% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data1.R
\docType{data}
\name{cakrill}
\alias{cakrill}
\title{Sample krill shape taken from krill in Monterey Bay (2019).}
\format{An object of class \code{FLS} of length 1.}
\usage{
cakrill

cakrill
}
\description{
Sample krill shape taken from krill in Monterey Bay (2019).

Sample krill shape taken from krill in Monterey Bay (2019).
}
\keyword{datasets}
