% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{CAL}
\alias{CAL}
\alias{CAL-class}
\title{Solid and calibration sphere (CAL) object/class.}
\value{
An object of the \code{CAL} S4 class.
}
\description{
A S4 class that provides slots to contain relevant metadata for solid sphere
objects belonging to the CAL-class scatterers. This object is created using
parameters specific to the outer shell. The default behavior of this object
is to only reference these outer elastic shell properties with few
exceptions that are model-dependent. \code{CAL} inherits from the broader
elastic-based \link{ELA} class rather than from \link{ESS}, because
calibration targets are solid elastic bodies rather than elastic shells. See
\link{Scatterer} for a more detailed description on how this S4 object is
organized.
}
\examples{
methods::getClass("CAL")

}
\seealso{
\link{Scatterer}
}
\keyword{scatterer_type}
