% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{cal_generate}
\alias{cal_generate}
\title{Generate a CAL-class object.}
\usage{
cal_generate(
  material = "WC",
  diameter = 0.0381,
  sound_speed_longitudinal = NULL,
  sound_speed_transversal = NULL,
  density_sphere = NULL,
  theta_sphere = pi,
  ID = NULL,
  diameter_units = "m",
  theta_units = "radians",
  n_segments = 100
)
}
\arguments{
\item{material}{Material type for the solid sphere. See 'Details' for
built-in material options.}

\item{diameter}{Spherical diameter (m).}

\item{sound_speed_longitudinal}{Longitudinal sound speed (m/s).}

\item{sound_speed_transversal}{Transversal sound speed (m/s).}

\item{density_sphere}{Density (kg/m^3).}

\item{theta_sphere}{Backscattering direction (Default: pi radians).}

\item{ID}{Optional metadata ID input.}

\item{diameter_units}{Compatibility argument. \code{cal_generate()} now assumes
meters and ignores non-SI alternatives.}

\item{theta_units}{Compatibility argument. \code{cal_generate()} now assumes
radians and ignores non-SI alternatives.}

\item{n_segments}{Number of segments to discretize object shape.}
}
\value{
Generates a CAL-class object.
}
\description{
Generate a CAL-class object.
}
\details{
There are several options for the \strong{material} argument:
\tabular{rlllll}{
\strong{Material} \tab \strong{Argument} \tab \strong{c1} \tab \strong{c2}
\tab \strong{\eqn{\rho1}}\cr
\emph{Tungsten carbide} \tab "WC" \tab 6853 \tab 4171 \tab 14900\cr
\emph{Stainless steel} \tab "steel" \tab 5980 \tab 3297 \tab 7970\cr
\emph{Brass} \tab "brass" \tab 4372 \tab 2100 \tab 8360\cr
\emph{Copper} \tab "Cu" \tab 4760 \tab 2288.5 \tab 8947\cr
\emph{Aluminum} \tab "Al" \tab 6260 \tab 3080 \tab 2700\cr
}
}
\examples{
cal_generate(material = "WC", diameter = 38.1e-3, n_segments = 120)

}
\seealso{
\code{\link{CAL}}
}
\keyword{scatterer_type_generation}
