% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cal_plot}
\alias{cal_plot}
\title{Plotting for CAL-class objects}
\usage{
cal_plot(
  object,
  type = "shape",
  nudge_y = 1,
  nudge_x = 1,
  x_units = "frequency",
  ...
)
}
\arguments{
\item{object}{CAL-class object.}

\item{type}{Toggle between body shape ("shape") or modeling results ("model")}

\item{nudge_y}{y-axis nudge.}

\item{nudge_x}{x-axis nudge.}

\item{x_units}{If "model" is selected, then toggle between frequency
("frequency", kHz) or ka ("ka").}

\item{...}{Additional plot inputs}
}
\description{
Plotting for CAL-class objects
}
