% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{cal_show}
\alias{cal_show}
\title{show(...) for CAL-class objects.}
\usage{
cal_show(object)
}
\arguments{
\item{object}{CAL-class object.}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
show(...) for CAL-class objects.
}
\examples{
cal_show(cal_generate(material = "WC", n_segments = 20))
}
\keyword{internal}
