% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{calculate_bessel_cache}
\alias{calculate_bessel_cache}
\title{Calculate Bessel function cache for the Goodman and Stern (1962) model}
\usage{
calculate_bessel_cache(ka_matrix_m, m)
}
\arguments{
\item{ka_matrix_m}{Modal ka matrix}

\item{m}{Modal vector}
}
\value{
Cached Bessel function values
}
\description{
Calculate Bessel function cache for the Goodman and Stern (1962) model
}
\keyword{internal}
