% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{calculate_goodman_stern_alpha}
\alias{calculate_goodman_stern_alpha}
\title{Calculate alpha coefficients for Goodman-Stern model}
\usage{
calculate_goodman_stern_alpha(
  bessel_cache,
  ka_matrix_m,
  m,
  lambda,
  mu,
  density_sw,
  density_shell,
  density_fluid
)
}
\arguments{
\item{bessel_cache}{Cached Bessel function values}

\item{ka_matrix_m}{Modal ka matrix}

\item{m}{Modal vector}

\item{lambda}{Lamé first parameter}

\item{mu}{Shear modulus}

\item{density_sw}{Seawater density}

\item{density_shell}{Shell density}

\item{density_fluid}{Fluid density}
}
\value{
List of alpha coefficients
}
\description{
Calculate alpha coefficients for Goodman-Stern model
}
