% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{calculate_goodman_stern_boundary_matrices}
\alias{calculate_goodman_stern_boundary_matrices}
\title{Calculate boundary condition matrices for Goodman and Stern (1962) model}
\usage{
calculate_goodman_stern_boundary_matrices(alpha, ka_matrix, m)
}
\arguments{
\item{alpha}{Alpha coefficient list}

\item{ka_matrix}{ka matrix}

\item{m}{Modal vector}
}
\value{
List of boundary matrices for each frequency and modal order
}
\description{
Calculate boundary condition matrices for Goodman and Stern (1962) model
}
