% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{calculate_ka_matrix}
\alias{calculate_ka_matrix}
\title{Calculate ka matrix for Goodman and Stern (1962) model}
\usage{
calculate_ka_matrix(
  frequency,
  sound_speed_sw,
  sound_speed_fluid,
  sound_speed_longitudinal,
  sound_speed_transversal,
  radius_shell,
  radius_fluid
)
}
\arguments{
\item{frequency}{Frequency vector}

\item{sound_speed_sw}{Seawater sound speed}

\item{sound_speed_fluid}{Fluid sound speed}

\item{sound_speed_longitudinal}{Longitudinal sound speed in shell}

\item{sound_speed_transversal}{Transversal sound speed in shell}

\item{radius_shell}{Shell radius}

\item{radius_fluid}{Fluid radius}
}
\value{
Matrix of ka values
}
\description{
Calculate ka matrix for Goodman and Stern (1962) model
}
