% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{CALgenerate}
\alias{CALgenerate}
\title{Generate CAL shape}
\usage{
CALgenerate(material = "WC", a = 0.0381, c1 = NULL, c2 = NULL, rho1 = NULL)
}
\arguments{
\item{material}{Material-type for the soldi sphere. See 'Details' built-in
material options.}

\item{a}{Spherical radius (m).}

\item{c1}{Longitudinal sound speed (m/s).}

\item{c2}{Transversal sound speed (m/s).}

\item{rho1}{Density (kg/m^3)}
}
\description{
Generate CAL shape
}
