% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calibration}
\alias{calibration}
\title{Calculates theoretical TS of a solid sphere of a certain material at a given
frequency.}
\usage{
calibration(object)
}
\arguments{
\item{object}{CAL-class object.}
}
\value{
The theoretical acoustic target strength (TS, dB re. 1 \eqn{m^2})  of a
solid sphere at a given frequency.
}
\description{
This function is a wrapper around TS_calculate(...) that parametrizes the
remainder of the model, while also doing simple calculations that do not
need to be looped. This function provides a TS estimate at a given frequency.
}
\references{
MacLennan D. N. (1981). The theory of solid spheres as sonar calibration
targets. Scottish Fisheries Research No. 22, Department of Agriculture and
Fisheries for Scotland.
}
