% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calibration_initialize}
\alias{calibration_initialize}
\title{Initialize CAL-class object for modeling.}
\usage{
calibration_initialize(
  object,
  frequency,
  sound_speed_sw = 1500,
  density_sw = 1026
)
}
\arguments{
\item{object}{CAL-class object.}

\item{frequency}{Frequency (Hz).}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}
}
\description{
Initialize CAL-class object for modeling.
}
