% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{create_shape}
\alias{create_shape}
\title{A wrapper function that automatically creates generalized and/or canonical
shapes for TS modeling.}
\usage{
create_shape(L, N, shape, ...)
}
\arguments{
\item{L}{Length or diameter of an object (m).}

\item{N}{Number of cylinders.}

\item{shape}{Shape. Details for shape specification are given under 'Details', 
including mandatory additional arguments.}

\item{g}{Optional density contrast input.}

\item{h}{Optional sound speed contrast input}

\item{theta}{Optional orientation input.}

\item{curve}{Optional shape curvature input [T/F].}

\item{pc}{Optional radius of curvature ratio input.}

\item{object}{Optional scatterer class object.}
}
\value{
Creates a generalized and/or canonical shape for TS modeling.
}
\description{
A wrapper function that automatically creates generalized and/or canonical
shapes for TS modeling.
}
\details{
The \strong{shape} argument specifies what shape for the function to generate into 
the desired shape for TS modeling. Options currently include:
\tabular{rlllll}{
 \tab \strong{Object Shape} \tab \strong{shape=...} \tab  \tab 
 \strong{Parameters} \tab \strong{Root function}\cr
 \tab \emph{Discrete Cylinder} \tab "cyl" \tab \tab LAratio \tab
 \code{\link[=cylinder]{cylinder(...)}}\cr
 \tab \emph{Tapered Cylinder} \tab "tcyl" \tab \tab LAratio, taper \tab 
 \code{\link[=tapered_cylinder]{tapered_cylinder(...)}}\cr
 \tab \emph{Polynomial Cylinder} \tab "pcyl" \tab \tab LAratio, poly \tab 
 \code{\link[=polynomial_cylinder]{polynomial_cylinder(...)}}\cr
 \tab \emph{Prolate Spheroid} \tab "psph" \tab \tab LAratio \tab 
 \code{\link[=prolate_spheroid]{prolate_spheroid(...)}}\cr 
 \tab \emph{Sphere} \tab "sph" \tab \tab None \tab 
 \code{\link[=sphere]{sphere(...)}}\cr
}

\subsection{Model Parameter Definitions}{
\itemize{
 \item \strong{LAratio}: the length-to-radius ratio (L/A), which specifically
 refers to the radius at the mid-point of the cylinder and should be 
 the maximum value. A typical L/A ratio in the literature is 16 for krill.
 \item \strong{taper}: the taper order (n), which parameterizes the tapering
 function reported by Chu \emph{et al.} (1993) to create a tapered cylinder.
 The tapering order will converge on a prolate and oblate spheroid when
 L > 2a and L < 2a, respectively, and n = 2. A typical taper order in the 
 literature is 10.
 \item \strong{poly}: the vector of arbitrary polynomial coefficients to 
 generate a deformed cylinder as reported by Smith \emph{et al.} (2013). 
 Although listed as a mandatory argument for the polynomial cylinder 
 function, it has a default setting that uses the sixth-degree polynomial 
 coefficients reported by Smith \emph{et al.} (2013). 
}
}
}
\references{
Chu, D., Foote, K.G., and Stanton, T.K. 1993. Further analysis of target 
strength measurements of Antarctic krill at 38 and 120 kHz: Comparison and 
deformed cylinder model and inference of orientation distribution. The 
Journal of the Acoustical Society of America, 93(5): 2985-2988. 
https://doi.org/10.1121/1.405818

Smith, J.N., Ressler, P.H., and Warren, J.D. 2013. A distorted wave Born 
approximation target strength model for Bering Sea euphausiids. ICES Journal
of Marine Science, 70(1): 204-214. https://doi.org/10.1093/icesjms/fss140
}
