% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{Cylinder}
\alias{Cylinder}
\alias{Cylinder-class}
\title{Cylindrical body shape}
\value{
An object of the \code{Cylinder} S4 class.
}
\description{
Cylindrical body shape
}
\examples{
methods::getClass("Cylinder")

}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
