% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{cylinder}
\alias{cylinder}
\title{Support function to generate tapered cylinder.}
\usage{
cylinder(
  L,
  LAratio,
  N,
  a = NULL,
  g = 1.0357,
  h = 1.0279,
  theta = pi/2,
  pc = 3,
  curve = F,
  object = "FLS"
)
}
\arguments{
\item{L}{Length of object (m).}

\item{LAratio}{Optional. Length:radius ratio.}

\item{N}{Number of cylinders.}

\item{a}{Maximum radius of object (m).}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Orientation.}

\item{pc}{Radius of curvature ratio.}

\item{curve}{Curvature [T/F].}

\item{object}{Object class.}
}
\value{
Creates a tapered cylinder.
}
\description{
Support function to generate tapered cylinder.
}
