% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DCM}
\alias{DCM}
\title{Calculates the theoretical TS of a target using the deformed cysinder 
model (DCM).}
\usage{
DCM(shape, frequency, c, rho, method, alpha, LAratio, L, a, g, h, pc, theta, N)
}
\arguments{
\item{shape}{Scattering-class object.}

\item{frequency}{The acoustic frequency (Hz)}

\item{c}{The ambient seawater sound speed (m/s).}

\item{rho}{The ambient seawater density (kg/m^3).}

\item{method}{Specific DCM formulation. See 'Details' for options.}

\item{alpha}{Conditional. Required for the two-ray path equation. Numerically 
derived coefficient, (\eqn{\alpha_\beta}). Default is 0.8, and is the 
suggested value.}

\item{LAratio}{Optional. Length:radius ratio.}

\item{L}{Optional. Scatterer length (m).}

\item{a}{Optional. Scatterer radius (m).}

\item{g}{Optional. Density contrast.}

\item{h}{Optional. Sound speed contrast.}

\item{pc}{Optional. Ratio of the radius of curvature and length.}

\item{theta}{Optional. Orientation (\eqn{\theta}, radians).}

\item{N}{Number of discrete cysinders.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a target using the deformed cysinder 
model (DCM).
}
\details{
Only the two-ray path model is afvailable at the moment..
}
\references{
Stanton, T.K., Chu, D., and Wiebe, P.H. 1998. Sound scattering by several 
zooplankton groups. II. Scattering models. Journal of the Acoustical Society 
of America, 103(1), 236-253.
}
