% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDWBA.R
\name{degrad}
\alias{degrad}
\title{Toggle between radians and degrees.}
\usage{
degrad(x,d)
}
\arguments{
\item{x}{A real value in degrees or radians}

\item{d}{The value input-type. Two input types: "deg" for degrees and "rad" for radians.}
}
\value{
Converts degrees to radians or radians to degrees
}
\description{
Toggle between radians and degrees.
}
\examples{
x <- 180 #degrees
degrad(x, "deg")
[1] 3.141593
x <- pi #radians
degrad(x, "rad")
[1] 180
}
