% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{degrees}
\alias{degrees}
\title{Convert angular measurements from radians to degrees}
\usage{
degrees(x)
}
\arguments{
\item{x}{A real value in radians}
}
\value{
Angle in degrees
}
\description{
Convert angular measurements from radians to degrees
}
\examples{
orientation <- pi / 2 # radians
degrees(orientation) # this should return a value equal to 90 degrees
}
