% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFCM.R
\name{DFCM}
\alias{DFCM}
\title{Calculates the theoretical TS of a fluid-filled scatterer at a given frequency using the deformed finite cylinder model (DFCM).}
\usage{
SDWBA(shape, f)
SDWBA(L, a, g, h, pc, numdiv, tilt, f)
SDWBA(shape, L, a, g, h, pc, numdiv, tilt, cm, pm, f, alpha, method)
}
\arguments{
\item{shape}{Desired object/animal shape. Must be class "FFS".}

\item{L}{Maximum length (m) of the scatterer. Can either be provided by shape input, or a manual value.}

\item{a}{Maximum radius (m) of the scatterer. Can either be provided by shape input, or a manual value.}

\item{g}{Density contrast (g) of the scatterer. Can either be provided by shape input, or a manual value.}

\item{h}{Sound speed contrast (h) of the scatterer. Can either be provided by shape input, or a manual value.}

\item{pc}{Radius of curvature (m) of the scatterer. Can either be provided by shape input, or a manual value.}

\item{numdiv}{Number of divisions to partition the animal shape into.
Can either be provided by the length of the shape position vector, or a manual value.}

\item{cm}{Sound speed of surrounding medium (m/s). Default value is 1500 m/s.}

\item{pm}{Density of surrounding medium (kg/m^3). Default value 1.025 kg/m^3.}

\item{f}{Frequency (Hz).}

\item{alpha}{Numerically derived coefficient (\eqn{\alpha_\beta}). Default is 0.8, and is the suggested value.}

\item{method}{Currently only the two-ray method is available, but the six-ray and other formulations will be available in the future.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a fluid-filled scatterer at a given frequency using the deformed finite cylinder model (DFCM).
}
\details{
Calculates the theoretical TS of a fluid-filled scatterer at a given frequency using the deformed finite cylinder model (DFCM).
}
\references{
Stanton, T.K., Chu, D., and Wiebe, P.H. 1998. Sound scattering by several zooplankton groups. II. Scattering models. Journal of the Acoustical Society of America, 103(1), 236-253.
}
