% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{discover_reforge_params}
\alias{discover_reforge_params}
\title{Get reforge parameters from known method signatures}
\usage{
discover_reforge_params(object_class)
}
\arguments{
\item{object_class}{Character string of object class}
}
\value{
Character vector of parameter names
}
\description{
Get reforge parameters from known method signatures
}
