% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.validate_brake_params}
\alias{.validate_brake_params}
\title{Validate brake/bending parameters}
\usage{
.validate_brake_params(body_df, radius_curvature, mode)
}
\description{
Validate brake/bending parameters
}
\keyword{internal}
