% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DWBA_curved}
\alias{DWBA_curved}
\title{Calculates the theoretical TS of a fluid-like scatterer at a given frequency
using the distorted Born wave approximation (DWBA) model.}
\usage{
DWBA_curved(object)
}
\arguments{
\item{object}{FLS-class scatterer.}
}
\description{
Calculates the theoretical TS of a fluid-like scatterer at a given frequency
using the distorted Born wave approximation (DWBA) model.
}
\references{
Stanton, T.K., Chu, D., and Wiebe, P.H. 1998. Sound scattering by several
zooplankton groups. II. Scattering models. J. Acoust. Soc. Am., 103, 236-253.

Demer, D.A., and Conti, S.G. 2003. Reconciling theoretical versus empirical
target strengths of krill: effects of phase variability on the distorted-wave
Born approximation. ICES J. Mar. Sci., 60: 429-434.
}
