% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{ELA}
\alias{ELA}
\alias{ELA-class}
\title{Elastic-based scatterer (ELA) object/class.}
\value{
An object of the \code{ELA} S4 class.
}
\description{
A shared parent S4 class for elastic-based scatterers. The \code{ELA}-class
collects the slots common to solid elastic targets such as calibration
spheres (\link{CAL}) and elastic-shelled targets (\link{ESS}), without
implying a particular internal structure such as a shell or homogeneous
elastic solid. See \link{Scatterer} for a more detailed description of the
shared object organization.
}
\examples{
methods::getClass("ELA")

}
\seealso{
\link{Scatterer}, \link{CAL}, \link{ESS}
}
\keyword{scatterer_type}
