% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{ESS}
\alias{ESS}
\alias{ESS-class}
\title{Elastic shelled scatterer (ESS) object/class.}
\value{
An object of the \code{ESS} S4 class.
}
\description{
A S4 class that provides slots to contain relevant metadata for elastic
shelled scatterers/objects belonging to the ESS-class. This object can be
created using values for both an outer shell and internal tissues, if
applicable. The default behavior for this type of this object is to only
reference the outer shell with few exceptions that are model-dependent.
\code{ESS} inherits from the broader elastic-based \link{ELA} class. See
\link{Scatterer} for a more detailed description on how this S4 object is
organized.
}
\examples{
methods::getClass("ESS")

}
\seealso{
\link{Scatterer}
}
\keyword{scatterer_type}
