% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{ESS-class}
\alias{ESS-class}
\alias{ESS}
\title{Elastic shelled scatterer (ESS) object/class.}
\description{
A S4 class that provides slots to contain relevant metadata for elastic
shelled scatterers/objects belonging to the ESS-class. This object can be
created using values for both an outer shell and internal tissues, if applicable.
The default behavior for this type of this object is to only reference the outer
shell with few exceptions that are model-dependent. See
\link{Scatterer} for a more detailed description on how
this S4 object is organized.
}
\seealso{
\link{Scatterer}
}
