% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{ess_show}
\alias{ess_show}
\title{show(...) for ESS-class objects.}
\usage{
ess_show(object)
}
\arguments{
\item{object}{ESS-class object.}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
show(...) for ESS-class objects.
}
\examples{
shell <- sphere(radius_body = 0.03, n_segments = 20)
target <- ess_generate(
  shape = shell,
  shell_thickness = 0.001,
  density_shell = 1050,
  sound_speed_shell = 2350,
  density_fluid = 1030,
  sound_speed_fluid = 1500,
  E = 3.5e9,
  nu = 0.34
)
ess_show(target)
}
\keyword{internal}
