% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{ESSgenerate}
\alias{ESSgenerate}
\title{Generate ESS shape}
\usage{
ESSgenerate(a, g, h, theta = pi/2, L = numeric(), spherical = TRUE)
}
\arguments{
\item{a}{Radius (m).}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Orientation (\eqn{\theta}, radians)}

\item{L}{Optional. Length of the longitudinal axis (m) for non-spherical 
shapes.}

\item{spherical}{Optional. Boolean for specifying whether scatterer is 
spherical or oblong.}
}
\description{
Generate ESS shape
}
