% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{extract,scatterer-method}
\alias{extract,scatterer-method}
\title{Method for what is printed for objects.}
\usage{
\S4method{extract}{scatterer}(object, feature)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{feature}{Feature of interest (e.g. body).}
}
\description{
Method for what is printed for objects.
}
