% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{fcms_initialize}
\alias{fcms_initialize}
\title{Initialize object for the modal series solution for a finite cylinder}
\usage{
fcms_initialize(
  object,
  frequency,
  boundary = "liquid_filled",
  sound_speed_sw = 1500,
  density_sw = 1026,
  m_limit = NULL
)
}
\description{
Initialize object for the modal series solution for a finite cylinder
}
