% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDWBA.R
\docType{class}
\name{FFS-class}
\alias{FFS-class}
\alias{FFS}
\title{Fluid-filled scatterer (FFS) object/class.}
\description{
A S4 class that provides slots to contain relevant animal metadata for parameterizing models for fluid-filled scatterers (FFS) partitioned
into discretized cylinders. This, specifically, includes a position matrix, radius, material properties (g, h), orientation,
animal shape, and body curvature. This class is used within the DWBA and DFCM model functions. In the future, this will also allow for
converting one class of scatterer into another for seemless usage for model comparisons.
}
