% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDWBA.R
\name{FFS_shape}
\alias{FFS_shape}
\title{Calls in a *.csv file as a FFS object}
\usage{
FFS_shape(file)
}
\arguments{
\item{file}{A *.csv file formatted with the following columns: x, y, z, a [radius], g, h}
}
\value{
Calls in an FFS_shape from a file
}
\description{
Calls in a *.csv file as a FFS object
}
