% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{FFSgenerate}
\alias{FFSgenerate}
\title{Manually generate a FFS object.}
\usage{
#Manually parameterize shape
x <- seq(1,10,1)*1e-3; y <- rep(0,10); z <- c(seq(1,5,1),rev(seq(1,5,1)))*1e-4
a <- z/2
g <- 1.036
h <- 1.0279
new_target <- FFSgenerate(x=x,y=y,z=z,a=a,g=g,h=h)
#Let's model where sound speed (c) is 1500 m/s, frequency is 120 kHz, with no phase deviation
c <- 1500
freq <- 120e3
SDWBA(shape=new_target, c=c, frequency=freq)
[1] -114.0107
}
\arguments{
\item{x}{Vector containing x-axis body (m) shape data.}

\item{y}{Vector containing y-axis body (m) shape data.}

\item{z}{Vector containing z-axis body (m) shape data.}

\item{a}{Vector containing radii (m).}

\item{g}{Density contrast.}

\item{h}{Soundspeed contrast}

\item{theta}{Orientation of the target relative to the transmit source (\eqn{\theta}). Broadside incidence is considered 90 degrees, or pi/2.
Default value is pi/2; input should be in radians.}
}
\value{
Calls in an FFS-class object from a *.csv file
}
\description{
Manually generate a FFS object.
}
