% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{FFSread}
\alias{FFSread}
\title{Calls in a *.csv file as a FFS object}
\usage{
FFSread(file)
}
\arguments{
\item{file}{A *.csv file formatted with the following columns: x, y, z, a, g, h, theta [optional], pc [optional]}
}
\value{
Calls in an FFS-class object from a *.csv file
}
\description{
Calls in a *.csv file as a FFS object
}
