% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{FLS-class}
\alias{FLS-class}
\alias{FLS}
\title{Fluid-like scatterer (FLS) object/class.}
\description{
A S4 class that provides slots to contain relevant metadata for scatterers
similar to the surrounding fluid medium (i.e fluid-like) belonging to
FLS-class scatterers. See \link[acousticTS]{scatterer} for a more detailed
description on how this S4 object is organized.
}
