% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_plot.R
\name{fls_plot}
\alias{fls_plot}
\title{Plotting for FLS-class objects}
\usage{
fls_plot(
  object,
  type = "shape",
  nudge_y = 1.05,
  nudge_x = 1.01,
  aspect_ratio = "manual",
  x_units = "frequency",
  y_units = "TS",
  ...
)
}
\arguments{
\item{object}{FLS-class object.}

\item{type}{Toggle between body shape ("shape") or modeling results ("model")}

\item{nudge_y}{y-axis nudge.}

\item{nudge_x}{x-axis nudge.}

\item{x_units}{If "model" is selected, then toggle between frequency
("frequency", kHz) or ka ("ka").}

\item{...}{Additional plot inputs}
}
\description{
Plotting for FLS-class objects
}
