% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{FLSread}
\alias{FLSread}
\title{Calls in a *.csv file as a FLS object}
\usage{
FLSread(file)
}
\arguments{
\item{file}{A *.csv file formatted with the following columns: x, y, z, a, g,
h, theta [optional], pc [optional]}
}
\value{
Calls in an FLS-class object from a *.csv file
}
\description{
Calls in a *.csv file as a FLS object
}
