% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{FLSwrite}
\alias{FLSwrite}
\title{Write FLS object as a .csv file.}
\usage{
FLSwrite(
  shape,
  filename = paste(getwd(), "/target_shape_", Sys.Date(), ".csv", sep = "")
)
}
\description{
Write FLS object as a .csv file.
}
