% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fsphere}
\alias{fsphere}
\title{Calculates the theoretical TS of a fluid sphere using an exact modal series 
solution proposed by Andersen (1950).}
\usage{
fsphere(shape, frequency, c)

fsphere(shape=NULL, frequency, c, a, g, h)
}
\arguments{
\item{shape}{Desired animal object. Must be class FS.}

\item{frequency}{Frequency (Hz).}

\item{c}{Sound speed of surrounding medium (m/s). Default value is 1500 m/s.}

\item{a}{Radius vector of sphere (m).}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{method}{Type of sphere being modeled. There are several options: 1 
(fluid-filled), 2 (fixed-rigid), 3 (pressure-release). Currently, fluid 
shells with fluid interiors are not available, but will be in future
versions.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a fluid sphere using an exact modal series 
solution proposed by Andersen (1950).
}
\details{
Calculates the theoretical TS of a fluid sphere at a given frequency using 
an exact modal series solution.
}
\references{
Anderson, V.C. (1950). Sound scattering from a fluid sphere. Journal of the 
Acoustical Society of America, 22, 426-431.
}
