% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{GAS}
\alias{GAS}
\alias{GAS-class}
\title{Generic gas-filled scatterer (GAS) object/class.}
\value{
An object of the \code{GAS} S4 class.
}
\description{
A S4 class that provides slots to contain relevant metadata for gas-bearing
scatterers belonging to the GAS-class. This object can include simple
gas-filled bubbles to other scatterers with gas occlusions, swimbladders,
and other internal features, if applicable. The default behavior for this
type of object is to only reference the gaseous/fluid feature with
exceptions that are model-dependent. See \link{Scatterer} for a more
detailed description on how this S4 object is organized.
}
\examples{
methods::getClass("GAS")

}
\seealso{
\code{\link{Scatterer}}
}
\keyword{scatterer_type}
