% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_plot.R
\name{gas_plot}
\alias{gas_plot}
\title{Plotting for GAS-class objects}
\usage{
gas_plot(
  object,
  type = "shape",
  nudge_y = 1.01,
  nudge_x = 1.01,
  x_units = "frequency",
  y_units = "TS",
  ...
)
}
\arguments{
\item{object}{GAS-class object.}

\item{type}{Toggle between body shape ("shape") or modeling results ("model")}

\item{nudge_y}{y-axis nudge.}

\item{nudge_x}{x-axis nudge.}

\item{x_units}{If "model" is selected, then toggle between frequency
("frequency", kHz) or ka ("ka").}

\item{y_units}{y-axis data selection (e.g. TS, sigma_bs -- defaults to TS).}

\item{...}{Additional plot inputs}
}
\description{
Plotting for GAS-class objects
}
