% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{gas_show}
\alias{gas_show}
\title{show(...) for GAS_class objects}
\usage{
gas_show(object)
}
\arguments{
\item{object}{GAS-class object}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
show(...) for GAS_class objects
}
\examples{
bubble <- gas_generate(
  shape = sphere(radius_body = 0.005, n_segments = 20)
)
gas_show(bubble)
}
\keyword{internal}
