% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%||\%}
\alias{\%||\%}
\title{Define null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Fallback value}
}
\value{
a if not NULL/NA, otherwise b
}
\description{
Returns first argument if not NULL and not NA, otherwise returns second.
}
\keyword{internal}
