% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{ha}
\alias{ha}
\alias{had}
\title{Bessel function of the third kind}
\usage{
ha(l, n)

had(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\description{
Bessel function of the third kind

First derivative of the Bessel function of the third kind
}
