% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{hcdk}
\alias{hcdk}
\title{k-th derivative of the cylindrical Hankel function of the first kind}
\usage{
hcdk(l, n, k)
}
\arguments{
\item{l}{A real integer or fractional order}

\item{n}{A real argument}

\item{k}{Order of the derivative (non-negative integer)}
}
\value{
Calculates the k-th derivative of the cylindrical Hankel function of the
first kind (\eqn{H_v^{(1)}}) with respect to its order.
}
\description{
k-th derivative of the cylindrical Hankel function of the first kind
}
\references{
DLMF 10.6.1, https://dlmf.nist.gov/10.6.E1
}
\keyword{bessel3}
