% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{high_pass_stanton}
\alias{high_pass_stanton}
\title{Calculates the theoretical TS of a shelled organism using the non-modal
High Pass (HP) model}
\usage{
high_pass_stanton(object)
}
\arguments{
\item{object}{Desired animal object (Elastic Shelled).}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a shelled organism using the non-modal
High Pass (HP) model
}
\references{
Lavery, A.C., Wiebe, P.H., Stanton, T.K., Lawson, G.L., Benfield, M.C.,
Copley, N. 2007. Determining dominant scatterers of sound in mixed
zooplankton popuilations. The Journal of the Acoustical Society of America,
122(6): 3304-3326.
}
