% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel-spherical.R
\name{hs}
\alias{hs}
\alias{hsdk}
\title{Spherical Bessel function of the third kind (Hankel), \eqn{h_\nu(x)}, and its
respective derivatives}
\usage{
hs(l, n)

hsdk(l, n, k)
}
\arguments{
\item{l}{Numeric. The order of the spherical Hankel function. Can be integer
or fractional.}

\item{n}{Numeric. The argument (\eqn{z}) at which to evaluate the function.}

\item{k}{Non-negative integer. The order of the derivative for \code{hsdk}.}
}
\value{
A complex vector containing:
\itemize{
\item \code{hs}: \eqn{h^{(1)}_\nu(z)}
\item \code{hsdK}: \eqn{\frac{d}{dz^k}h^{(1)}_l(z)} (k-th derivative)
}
}
\description{
Computes the spherical Hankel function of the first kind
(\eqn{h^{(1)}_\nu(z)}) and its first-th derivative.
}
\details{
The spherical Hankel function of the first kind is defined as:
\deqn{h^{(1)}_\nu(z) = j_\nu(z) + i y_\nu(z)}

where \eqn{j_\nu(z)} is the spherical Bessel function of the first kind and
\eqn{y_\nu(z)} is the spherical Bessel function of the second kind.

It is related to the cylindrical Hankel function by:
\deqn{h^{(1)}_\nu(z) = \sqrt{\frac{\pi}{2z}} H^{(1)}_{\nu+1/2}(z)}

The spherical Hankel functions are used extensively in scattering theory
to represent outgoing spherical waves.

\strong{Derivative:}
\deqn{
 \frac{d}{dz}h^{(1)}_\nu(z) = \frac{\nu}{z} h^{(1)}_\nu(z)
 - h^{(1)}_{\nu+1}(z)
}
}
\examples{
# Spherical Hankel function
hs(0, 1)
hs(1, 2.5)

# Fractional order
hs(0.5, 3)

# Vector input
hs(0, c(1, 2, 3))

# First derivative
hsdk(1, 2, 1)

}
\references{
Abramowitz, M. and Stegun, I.A. (Eds.). (1964). \emph{Handbook of
Mathematical Functions with Formulas, Graphs, and Mathematical Tables}.
National Bureau of Standards, Applied Mathematics Series 55. Chapter 10.

NIST Digital Library of Mathematical Functions. \url{https://dlmf.nist.gov/}
\itemize{
\item Spherical Bessel functions: \url{https://dlmf.nist.gov/10.47}
\item Relation to cylindrical Hankel functions: Eq. 10.47.5 at
\url{https://dlmf.nist.gov/10.47}
}
}
\seealso{
\code{\link{hc}} for cylindrical Hankel functions,
\code{\link{js}} for spherical Bessel functions of the first kind,
\code{\link{ys}} for spherical Bessel functions of the second kind.
}
\keyword{bessel}
