% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R, R/calibration_sphere.R
\name{ja}
\alias{ja}
\title{Calculate Bessel function of the first kind.}
\usage{
ja(l,n)

ja(l,n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}

\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Bessel function of the first kind (J_a).

Calculates the Bessel function of the first kind (J_a).
}
\description{
Calculate Bessel function of the first kind.

Calculate Bessel function of the first kind.
}
\examples{
l <- 1.5
n <- 1
ja(l,n)
[1] 0.2402978
l <- 1.5
n <- 1
ja(l,n)
[1] 0.2402978
}
\references{
Amost D.E., "AMOS, A Portable Package for Bessel Functions of a Complex Argument and Nonnegative Order", http://netlib.org/amos

Amost D.E., "AMOS, A Portable Package for Bessel Functions of a Complex Argument and Nonnegative Order", http://netlib.org/amos
}
