% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{jl1}
\alias{jl1}
\title{Calculate spherical Bessel functions.}
\usage{
jl(l,n,sign) 
yl(l,n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}

\item{sign}{Flag used to to determine whether to add or subtract 0.5 from the function's order (l); only used 
for spherical Bessel function of first kind.}
}
\value{
jl(l,n,sign) calculates the spherical Bessel function of the first kind (zbesj AMOS routine, Iv).
yl(l,n) calculates the spherical Bessel function of the second kind (zbesy AMOS routine, Yv)
}
\description{
These functions calculate spherical Bessel functions of first and second kinds using the scipy.special module from Python.
Functions are referenced from Amost D.E., "AMOS, A Portable Package for Bessel Functions of a Complex Argument
and Nonnegative Order", http://netlib.org/amos
}
\examples{
l <- 1
n < 1 
jl(l,n,sign=1)
[1] 0.3011687
yl(l,n)
[1] -1.381773
}
