% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{k}
\alias{k}
\title{Calculate the acoustic wavenumber (k) based on the sound speed of water.}
\usage{
k(frequency, sound_speed)
}
\arguments{
\item{frequency}{Frequency (f, Hz)}

\item{sound_speed}{Sound speed (c, m/s)}
}
\value{
Calculates the acoustic wavenumber (k) based on the sound speed of water.
}
\description{
Calculate the acoustic wavenumber (k) based on the sound speed of water.
}
