% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{kappa}
\alias{kappa}
\title{Calculate the compressibility material properties of a scatterer's tissue or
shell (kappa)}
\usage{
kappa(interface1, interface2)
}
\arguments{
\item{interface1}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (1)}

\item{interface2}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (2)}
}
\description{
Calculate the compressibility material properties of a scatterer's tissue or
shell (kappa)
}
