% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{kcalc}
\alias{kcalc}
\title{Calculate the acoustic wavenumber based on the sound speed of water.}
\usage{
kcalc(frequency, sound_speed)
}
\arguments{
\item{frequency}{Frequency (f, Hz)}

\item{sound_speed}{Sound speed (c, m/s)}
}
\value{
Calculates the acoustic wavenumber based on the sound speed of water
}
\description{
Calculate the acoustic wavenumber based on the sound speed of water.
}
\examples{
c <- 1500 #m/s
f <- 120e3 #Hz
kcalc(f,c)
# 502.6547
}
