% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{krill}
\alias{krill}
\title{Sample krill (Euphausia superba) shape taken from McGehee et al. (1998)}
\format{
A pre-generated FLS scatterer containing all information required
for target strength modeling.
\describe{
\item{metadata}{Relevant and identifying metadata (\code{list}).}
\item{model_parameters}{Container for specified model parameters
(\code{list}).}
\item{model}{Model outputs and results (\code{list}).}
\item{body}{A list with elements:
\itemize{
\item \code{rpos}: Position matrix (x, y, z; m).
\item \code{radius}: Radius of each discrete cylinder (m).
\item \code{g}: Body density contrast relative to medium.
\item \code{h}: Sound speed contrast relative to medium.
\item \code{theta}: Orientation angle (\eqn{\theta_{body}}, radians).
}
}
\item{shape_parameters}{A list with:
\itemize{
\item \code{length}: Body length (m).
\item \code{ncyl}: Number of cylinders.
\item \code{theta_units}: Unit of orientation.
\item \code{length_units}: Unit of length.
}
}
}
}
\usage{
data(krill)
}
\value{
A pre-generated \code{FLS} scatterer object.
}
\description{
A dataset containing a sample krill (Euphausia superba) body shape proposed
by McGehee et al. (1998).
}
\examples{
data("krill", package = "acousticTS")
krill
}
\keyword{datasets}
