% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRM.R
\name{KRM.sim}
\alias{KRM.sim}
\title{Wrapper function to simulate KRM runs.}
\usage{
KRM.sim(
  shape = shape,
  c = 1500,
  rho = 1030,
  frequency,
  theta = shape@theta,
  pb = shape@pb,
  cb = shape@cb,
  length = shape@L,
  permute = T,
  progress = T,
  parallel = F
)
}
\description{
Wrapper function to simulate KRM runs.
}
