% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{krm_initialize}
\alias{krm_initialize}
\title{Initialize SBF-class object for KRM calculations.}
\usage{
krm_initialize(
  object,
  frequency,
  sound_speed_sw = 1500,
  density_sw = 1026,
  density_body = NULL,
  density_swimbladder = NULL,
  sound_speed_body = NULL,
  sound_speed_swimbladder = NULL,
  theta = NULL,
  body_resize = 1,
  swimbladder_resize = 1
)
}
\arguments{
\item{object}{SBF-class object}

\item{frequency}{Frequency (Hz).}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}

\item{density_body}{Optional flesh density input.}

\item{density_swimbladder}{Optional gas density input.}

\item{sound_speed_body}{Optional flesh sound speed input.}

\item{sound_speed_swimbladder}{Optional gas sound speed input.}

\item{theta}{Optional orientation input (relative to incident sound wave).}

\item{body_resize}{Factor resize for body.}

\item{swimbladder_resize}{Factor resize for swimbladder.}
}
\description{
Initialize SBF-class object for KRM calculations.
}
