% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{lame}
\alias{lame}
\title{Calculate Lamé's first parameter (\eqn{\lambda})}
\usage{
lame(K = NULL, E = NULL, G = NULL, nu = NULL)
}
\arguments{
\item{K}{Bulk modulus (Pa).}

\item{E}{Young's modulus (Pa).}

\item{G}{Shear modulus (Pa).}

\item{nu}{Poisson's ratio (Dimensionless).}
}
\value{
Lamé's first parameter (\eqn{\lambda}, Pa).
}
\description{
Calculates Lamé's first parameter (\eqn{\lambda}) from two of the
four other elastic moduli: bulk modulus (K), Young's modulus (E), shear
modulus (G), or Poisson's ratio (\eqn{\nu}). Assumes 3D material properties.

The relationships used are:
\deqn{\lambda = K - \frac{2G}{3}}
\deqn{\lambda = \frac{E\nu}{(1 + \nu)(1 - 2\nu)}}
\deqn{\lambda = \frac{2G\nu}{1 - 2\nu}}
\deqn{\lambda = \frac{3K\nu}{1 + \nu}}
\deqn{\lambda = \frac{3K(3K - E)}{9K - E}}
\deqn{\lambda = \frac{G(E - 2G)}{3G - E}}
}
\examples{
lame(K = 2.5e9, G = 1.1e9)

}
\keyword{elastic}
