% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcgehee}
\alias{mcgehee}
\title{Sample krill shape taken from McGehee et al. (1998)}
\format{
A pre-generated SBF scatterer containing all information required for
target strength modeling.
\describe{
\item{metadata}{Relevant and identifying metadata (\code{list}).}
\item{model_parameters}{Container for specified model parameters (\code{list}).}
\item{model}{Model outputs and results (\code{list}).}
\item{body}{
\itemize{
\item{rpos:}{ Position matrix (x, y, z; m).}
\item{radius:}{ Radius of each discrete cylinder along the position
matrix (m).}
\item{g:}{ Body density contrast relative to the surrounding medium.}
\item{h:}{ Sound speed contrast relative to the surrounding medium.}
\item{theta:}{ Body orientation relative to transmitting transducer (\ifelse{html}{\out{&theta;<sub>body</sub>}}{\eqn{\theta_{body}}}
, rads)}}}
\item{shape_parameters}{
\itemize{
\item{body}{
\itemize{
\item{length:}{ Body length (m).}
\item{ncyl:}{ Number of discrete cylinders along body length.}
\item{theta_units:}{ Units used for orientation of body.}
\item{length_units:}{ Units used for body length.}}
}}
}}
}
\usage{
data(mcgehee)
}
\description{
A dataset containing a sample krill (Euphausia superba) body shape proposed
by McGehee et al. (1998).
}
\keyword{datasets}
