% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modal_matrix}
\alias{modal_matrix}
\title{Format data for the modal series solution model into the appropriate matrix}
\usage{
modal_matrix(v, limit)
}
\arguments{
\item{v}{Vector input.}

\item{limit}{Modal series limit.}
}
\description{
Format data for the modal series solution model into the appropriate matrix
}
\keyword{internal}
