% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_plot.R
\docType{data}
\name{model_palette}
\alias{model_palette}
\title{Base plotting color palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
model_palette
}
\description{
Color palette vector referenced for plotting multiple models simultaneously.
}
\keyword{datasets}
