% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{model_registry}
\alias{model_registry}
\title{Model registry: maps model names to their functions and related initializers}
\format{
An object of class \code{list} of length 10.
}
\usage{
model_registry
}
\description{
Model registry: maps model names to their functions and related initializers
}
\keyword{datasets}
