% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{MSS_anderson}
\alias{MSS_anderson}
\title{Calculates the theoretical TS of a fluid sphere using an exact modal series
solution proposed by Andersen (1950).}
\usage{
MSS_anderson(object)
}
\arguments{
\item{object}{GAS- or SBF-class object.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a fluid sphere using an exact modal series
solution proposed by Andersen (1950).
}
\details{
Calculates the theoretical TS of a fluid sphere at a given frequency using
an exact modal series solution.
}
\references{
Anderson, V.C. (1950). Sound scattering from a fluid sphere. Journal of the
Acoustical Society of America, 22, 426-431.
}
