% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{MSS_goodman_stern}
\alias{MSS_goodman_stern}
\title{Calculates the theoretical TS of an elastic-shelled sphere using the modal
series solution from Goodman and Stern (1962).}
\usage{
MSS_goodman_stern(object)
}
\arguments{
\item{object}{GAS- or SBF-class object.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of an elastic-shelled sphere using the modal
series solution from Goodman and Stern (1962).
}
\details{
Calculates the theoretical TS of an elastic-shelled sphere using an exact
modal series solution
}
\references{
Goodman, R.R., and Stern, R. (1962). Reflection and transmission of sound by
elastic spherical shells. The Journal of the Acoustical Society of America,
34, 338-344.
}
