% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{mss_goodman_stern_initialize}
\alias{mss_goodman_stern_initialize}
\title{Initialize ESS-object for modal series solution.}
\usage{
mss_goodman_stern_initialize(
  object,
  frequency,
  sound_speed_sw = 1500,
  density_sw = 1026,
  m_limit = NULL
)
}
\arguments{
\item{object}{ESS-class object.}

\item{frequency}{Frequency vector (Hz).}

\item{sound_speed_sw}{Seawater sound speed (m/s). Default: 1500.}

\item{density_sw}{Seawater density (kg/m³). Default: 1026.}

\item{m_limit}{Modal series limit (i.e. max "m"). The default is the maximum
ka + 10.}
}
\description{
Initialize ESS-object for modal series solution.
}
