% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{neumann}
\alias{neumann}
\title{Compute the Neumann factor \eqn{\nu_{n}}}
\usage{
neumann(x)
}
\arguments{
\item{x}{An integer iterator.}
}
\value{
A numeric vector of the same length as \code{x}, containing values of
\eqn{\eta_x}, each equal to 1 or 2.
}
\description{
The Neumann factor, denoted \eqn{\nu_{n}}, is a simple multiplicative
constant commonly used in spherical or spheroidal function theory. It
accounts for the symmetry of cosine terms or the duplication of even-order
contributions in integrals.
}
\details{
Formally, the Neumann factor is defined as:
\deqn{
  \eta_n = \begin{cases}
    1, & n = 0, \\
    2, & n > 0.
  \end{cases}
}

This factor frequently appears in the normalization of spherical Bessel
functions, Legendre expansions, and spheroidal wave functions. It is not
related to the "von Neumann ordinals" used in set theory.
}
\examples{
neumann(0) # should return 1
neumann(1) # should return 2
neumann(2) # should return 2

# Vectorized use:
neumann(0:4)

}
