% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{oblate_spheroid}
\alias{oblate_spheroid}
\title{Creates an oblate spheroid.}
\usage{
oblate_spheroid(
  length_body = NULL,
  radius_body = NULL,
  length_radius_ratio = NULL,
  semimajor_length = NULL,
  semiminor_length = NULL,
  n_segments = 100,
  length_units = "m"
)
}
\arguments{
\item{length_body}{Body-axis length (m).}

\item{radius_body}{Maximum equatorial radius (m).}

\item{length_radius_ratio}{Optional ratio input when radius is not explicitly
known.}

\item{semimajor_length}{Optional alias for maximum equatorial radius (m).}

\item{semiminor_length}{Optional alias for body-axis semi-length (m).}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
100 segments.}

\item{length_units}{Units for body matrix (defaults to m).}
}
\value{
Creates the position vector for an oblate spheroid object of defined
body-axis length and maximum equatorial radius.
}
\description{
Creates an oblate spheroid.
}
\examples{
oblate_spheroid(
  length_body = 0.012, radius_body = 0.01, n_segments = 60
)
oblate_spheroid(
  semiminor_length = 0.006, semimajor_length = 0.01
)
}
\seealso{
\code{\link{OblateSpheroid}}
}
\keyword{shape_generation}
