% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{OblateSpheroid}
\alias{OblateSpheroid}
\alias{OblateSpheroid-class}
\title{Oblate spheroidal body shape}
\value{
An object of the \code{OblateSpheroid} S4 class.
}
\description{
Oblate spheroidal body shape
}
\examples{
methods::getClass("OblateSpheroid")

}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
