% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{phase_integrate}
\alias{phase_integrate}
\title{Wrapper function incorporating phase deviation into contour integration}
\usage{
phase_integrate(x, y, n_iterations, integral)
}
\arguments{
\item{x}{Indexing argument for multi-row objects}

\item{y}{Indexing argument for multi-column objects}

\item{n_iterations}{Number of phase deviations to average and summarize}

\item{integral}{Integral function used for numerical integration via adaptive
quadrature}
}
\description{
Wrapper function incorporating phase deviation into contour integration
}
