% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,scatterer,missing-method}
\alias{plot,scatterer,missing-method}
\title{Method for what is printed for objects.}
\usage{
\S4method{plot}{scatterer,missing}(
  x,
  type = "shape",
  nudge_y = 1.1,
  nudge_x = 1.05,
  x_units = "frequency",
  ...
)
}
\arguments{
\item{x}{Scatterer-class object.}

\item{type}{Toggle between body shape ("shape") or modeling results ("model")}

\item{nudge_y}{y-axis nudge.}

\item{nudge_x}{x-axis nudge.}

\item{x_units}{If "model" is selected, then toggle between frequency
("frequency", kHz) or ka ("ka").}

\item{...}{Additional plot inputs}
}
\description{
Method for what is printed for objects.
}
