% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot,scattering,missing-method}
\alias{plot,scattering,missing-method}
\title{Method for plotting scatterer-class objects.}
\usage{
\S4method{plot}{scattering,missing}(x, nudge_y = 1, ...)
}
\description{
Method for plotting scatterer-class objects.
}
