% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_plot.R
\docType{data}
\name{plot_colors}
\alias{plot_colors}
\title{Base plotting color palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
plot_colors
}
\description{
Base plotting color palette
}
\keyword{datasets}
