% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{polynomial_cylinder}
\alias{polynomial_cylinder}
\title{Creates a polynomial deformed cylinder.}
\usage{
polynomial_cylinder(length, radius, n_segments, polynomial)
}
\arguments{
\item{n_segments}{Number of segments to discretize object shape. Defaults to
1e2 segments.}

\item{polynomial}{Polynomial coefficient vector.}
}
\value{
Creates the position vector for a polynomial deformed cylinder.
}
\description{
Creates a polynomial deformed cylinder.
}
\examples{
\dontrun{
# We can use the polynomial coefficients defined in Smith et al. (2013) to
# define the position vector of a sub-Arctic krill.
poly_vec <- c(0.83, 0.36, -2.10, -1.20, 0.63, 0.82, 0.64)
# Create the position vector
# This outputs a list containing "rpos" and "radius"
pos <- polynomial_cylinder(length = 15e-3, radius = 2e-3, polynomial = poly_vec)
str(pos)
}
}
\references{
Smith, J.N., Ressler, P.H., and Warren, J.D. 2013. A distorted wave Born
approximation target strength model for Bering Sea euphausiids. ICES Journal
of Marine Science, 70(1): 204-214. https://doi.org/10.1093/icesjms/fss140
}
