% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{PolynomialCylinder}
\alias{PolynomialCylinder}
\alias{PolynomialCylinder-class}
\title{Cylindrical body shape deformed using a polynomial}
\value{
An object of the \code{PolynomialCylinder} S4 class.
}
\description{
Cylindrical body shape deformed using a polynomial
}
\examples{
methods::getClass("PolynomialCylinder")

}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
