% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pos_matrix}
\alias{pos_matrix}
\title{Extract position matrix from object for plotting and other purposes.}
\usage{
pos_matrix(object)
}
\description{
Extract position matrix from object for plotting and other purposes.
}
