% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{prolate_spheroid}
\alias{prolate_spheroid}
\title{Creates a prolate spheroid.}
\usage{
prolate_spheroid(
  length_body = NULL,
  radius_body = NULL,
  length_radius_ratio = NULL,
  semimajor_length = NULL,
  semiminor_length = NULL,
  n_segments = 100,
  length_units = "m"
)
}
\arguments{
\item{length_body}{Semi-major axis length (m).}

\item{radius_body}{Semi-minor axis length (m). This can also be stylized as
the "maximum radius" of the scattering object.}

\item{length_radius_ratio}{Optional ratio input when radius is not explicitly
known.}

\item{semimajor_length}{Optional alias for semi-major axis length (m).}

\item{semiminor_length}{Optional alias for semi-minor axis length (m).}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
18 segments.}

\item{length_units}{Units for body matrix (defaults to m).}
}
\value{
Creates the position vector for a prolate spheroid object of defined
semi-major and -minor axes.
}
\description{
Creates a prolate spheroid.
}
\examples{
prolate_spheroid(
  length_body = 0.04, radius_body = 0.004, n_segments = 60
)
prolate_spheroid(
  semimajor_length = 0.05, semiminor_length = 0.003
)
}
\seealso{
\code{\link{ProlateSpheroid}}
}
\keyword{shape_generation}
