% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{ProlateSpheroid}
\alias{ProlateSpheroid}
\alias{ProlateSpheroid-class}
\title{Prolate spheroidal body shape}
\value{
An object of the \code{ProlateSpheroid} S4 class.
}
\description{
Prolate spheroidal body shape
}
\examples{
methods::getClass("ProlateSpheroid")

}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
