% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{psms_initialize}
\alias{psms_initialize}
\title{Initialize object for the modal series solution for a prolate spheroid}
\usage{
psms_initialize(
  object,
  frequency,
  phi_body = pi,
  boundary = "liquid_filled",
  simplify_Amn = FALSE,
  sound_speed_sw = 1500,
  density_sw = 1026
)
}
\description{
Initialize object for the modal series solution for a prolate spheroid
}
