% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonics.R
\name{Qndk}
\alias{Qndk}
\title{Derivative of the Legendre function of the second kind}
\usage{
Qndk(n, x, k = 1L)
}
\arguments{
\item{n}{Numeric vector. Degree (order) of the Legendre function.}

\item{x}{Numeric vector. Real argument(s) at which to evaluate the
derivative. Avoid \eqn{x = \pm 1}.}

\item{k}{Integer. Order of the derivative (\eqn{k \geq 0}). Default is 1.}
}
\value{
A complex matrix of dimension \code{length(n)} by \code{length(x)}, where
element \code{[i, j]} contains \eqn{\frac{d^k}{dx^k} Q_{n_i}(x_j)}.
}
\description{
Derivative of the Legendre function of the second kind
}
\note{
Derivatives are computed via finite differences with step size
\eqn{h = 10^{-6}}. Accuracy is typically 4-6 significant digits for
first derivatives, less for higher orders.

This function calls underlying \eqn{C++} code via \code{Rcpp} for
computational efficiency and to support different cases for both order and
argument that are not readily available in \code{R}.
}
\examples{
# First derivative of Q_1(x) at x = 0.5
Qndk(1, 0.5, 1)

# Compare with numerical derivative
h <- 1e-6
(Qn(1, 0.5 + h) - Qn(1, 0.5 - h)) / (2 * h)

# Second derivative
Qndk(2, 0.5, 2)

# Multiple orders
Qndk(c(1, 2, 3), 0.5, 1)

# Complex result for |x| > 1
Qndk(1, 2.0, 1)

}
\references{
Abramowitz, M. and Stegun, I. A. (1972). \emph{Handbook of Mathematical
Functions}. Dover Publications. Chapter 8: Legendre Functions.
}
\seealso{
\code{\link{Qn}} for Legendre functions of the second kind.
}
\keyword{derivative}
\keyword{legendre}
