% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{R12}
\alias{R12}
\title{Calculate reflection coefficient}
\usage{
R12(body, medium)
}
\arguments{
\item{body}{Body material properties (i.e. g, h)}

\item{medium}{Seawater material properties (i.e., density, sound speed)}
}
\description{
Calculate reflection coefficient
}
