% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{radians}
\alias{radians}
\alias{degrees}
\title{Convert between degrees and radians}
\usage{
radians( x )
degrees( x )

degrees(x)
}
\arguments{
\item{x}{A real value in degrees or radians
for radians.}
}
\value{
Converts degrees to radians or radians to degrees
}
\description{
Convert between degrees and radians
}
