% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{reflection_coefficient}
\alias{reflection_coefficient}
\title{Calculate reflection coefficient}
\usage{
reflection_coefficient(material_properties_1, material_properties_2)
}
\arguments{
\item{material_properties_1}{Interface 1}

\item{material_properties_2}{Interface 2}
}
\description{
Calculate reflection coefficient
}
