% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge,CAL-method}
\alias{reforge,CAL-method}
\title{Reforge CAL-class object}
\usage{
\S4method{reforge}{CAL}(object, scale = NULL, diameter_target = NULL, n_segments = NULL)
}
\arguments{
\item{object}{CAL-class object.}

\item{scale}{Single positive scale factor applied isometrically. Mutually
exclusive with \code{diameter_target}.}

\item{diameter_target}{Target sphere diameter (m). Derives the scale factor
internally. Mutually exclusive with \code{scale}.}

\item{n_segments}{New number of discrete segments along the major axis.}
}
\value{
Modified CAL-class object.
}
\description{
Resize a calibration sphere by applying an isometric scale factor or
specifying a target diameter.  Optionally re-discretize to a new segment
count.  CAL objects are always spheres, so the position matrix follows
the \code{sphere()} convention (n_points x 5: x, y, z, zU, zL).
}
\examples{
methods::selectMethod("reforge", "CAL")
}
\keyword{internal}
