% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resize}
\alias{resize}
\title{Resize animal to maintain shape based on length.}
\usage{
resize(shape, length)
}
\arguments{
\item{shape}{Desired object/animal shape.}

\item{length}{New length (m).}
}
\value{
Rescales the shape of an animal based on a desired length.
}
\description{
Resize animal to maintain shape based on length.
}
