% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{rho}
\alias{rho}
\title{Calculate the density contrast (\eqn{\rho}) of a scattering boundary}
\usage{
rho(medium, target)
}
\arguments{
\item{medium}{Dataframe object containing density (\eqn{kgm^{-3}}) and
sound speed (\eqn{ms^{-1}}) values for the external medium.}

\item{target}{Dataframe object containing density (\eqn{kgm^{-3}}) and
sound speed (\eqn{ms^{-1}}) values for the target boundary.}
}
\value{
Density contrast, defined as
\eqn{(\rho_{target} - \rho_{medium})/\rho_{target}}.
}
\description{
Calculate the density contrast (\eqn{\rho}) of a scattering boundary
}
\examples{
seawater <- data.frame(density = 1026, sound_speed = 1480)
animal <- data.frame(density = 1050, sound_speed = 1530)
rho(seawater, animal)
}
