% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{SBF}
\alias{SBF}
\alias{SBF-class}
\title{Swimbladdered fish (SBF) object/class.}
\value{
An object of the \code{SBF} S4 class.
}
\description{
A S4 class that provides slots to contain relevant animal metadata for
parameterizing models for swimbladdered fish (SBF) that are partitioned into
two sets of discretized cylinders: the body and the swimbladder. Both shapes
comprise independent position matrices, material properties, orientations,
and other relevant shape-related data and metadata. See
\link{Scatterer} for a more detailed description on how
this S4 object is organized.
}
\examples{
methods::getClass("SBF")

}
\seealso{
\link{Scatterer}
}
\keyword{scatterer_type}
