% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{SBF-class}
\alias{SBF-class}
\alias{SBF}
\title{Swimbladdered fish (SBF) object/class.}
\description{
A S4 class that provides slots to contain relevant animal metadata for
parameterizing models for swimbladdered fish (SBF) that are partitioned into
two sets of discretized cylinders: the body and the swimbladder. Both shapes
comprise independent position matrices, material properties, orientations,
and other relevant shape-related data and metadata.See \link[acousticTS]{scatterer}
for a more detailed description on how this S4 object is organized.
}
