% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{SBFgenerate}
\alias{SBFgenerate}
\title{Generate SBF shape}
\usage{
SBFgenerate(
  xb,
  wb,
  zbU,
  zbL,
  xsb,
  wsb,
  zsbU,
  zsbL,
  pb,
  cb,
  psb,
  csb,
  theta = pi/2
)
}
\description{
Generate SBF shape
}
