% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{SBFread}
\alias{SBFread}
\title{Calls in a *.csv file as a SBF}
\usage{
SBFread(file)
}
\arguments{
\item{file}{A *.csv file formatted with the following columns: xb, wb, zbU, 
zbL, xsb, wsb, zsbU, zsbL, pb, cb, psb, csb, theta [optional].}
}
\value{
Calls in an SBF-class object from a *.csv file
}
\description{
Calls in a *.csv file as a SBF
}
