% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{scattering-class}
\alias{scattering-class}
\alias{scattering}
\title{Scattering-class objects.}
\description{
The group of S4 class objects that represents different scatterer-types, such
as fluid-like (FLS) and gas-bearing (GAS) animals. In general, each of these 
classes contain relevant animal metadata for parameterizing models, such 
as orientation (\eqn{\theta}), material properties (\eqn{\rho} and g, 
c and h), and body morphology (e.g., \eqn{r_0}).
}
\details{
There are currently several scattering-classes available for modeling using
the \eqn{acousticTS} package:

\itemize{
 \item \strong{FLS}: \code{\link[acousticTS; classes; FLS]{fluid-like 
 scatterers (FLS)}} defined 
 by the FLS-class. These generally describe scatterers (or body structures) 
 that are fluid-like, meaning that the acoustic material properties (i.e., 
 sound speed and density) are similar to that of the surrounding medium. 
 Consequently, things like thin flesh and crustaceans are typically modeled 
 as FLS objects. Current models available to FLS objects includes the 
 \link[acousticTS; models]{SDWBA}, \link[acousticTS; models]{KRM}, and 
 \link[acousticTS; models]{DCM}.
 \item \strong{SBF}: `swimbladdered fish (SBF)` defined by the SBF-class. 
 \item \strong{CAL}: `solid and calibration spheres (CAL)` defined by the 
 CAL-class. 
 \item \strong{GAS}: `gas-bearing scatterers (GAS)` defined by the GAS-class.
 \item \strong{ESS}: `elastic shelled scatterers (ESS)` defined by the 
 ESS-class.
}
}
