% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDWBA.R
\name{SDWBA.sim}
\alias{SDWBA.sim}
\title{Wrapper function that can simulate over distributions of values}
\usage{
SDWBA.sim(shape, c, f, phi, theta)
}
\arguments{
\item{shape}{Desired object/animal shape. Must be class "FFS".}

\item{c}{Sound speed of surrounding medium (m/s). Default value is 1500 m/s.}

\item{theta}{Orientation of the target relative to the transmit source (\eqn{\theta}). Broadside incidence is considered 90 degrees, or pi/2.
Default value is pi/2; input should be in radians.}

\item{nrep}{Number of times/iterations the model will be ran.}

\item{aggregate}{Will output an aggregate statistic with the following options: "mean", "median", "maximum", and "minimum". Multiple, or all, of these options are also available.}

\item{f}{Frequency (Hz).}

\item{phi}{Phase deviation (\eqn{\phi}), or phase variability. Accounts for complexities in animal shape and stochasticity of noise in scattering field.
Default value is 0.0.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Wrapper function that can simulate over distributions of values
}
\details{
Calculates the theoretical TS of a fluid-filled scatterer at a given frequency using the distorted Born wave approximation (DWBA) model.
}
\references{
Stanton, T.K., Chu, D., and Wiebe, P.H. 1998. Sound scattering by several zooplankton groups. II. Scattering models. Journal of the Acoustical Society of America, 103(1), 236-253.

Hankin, R.K.S. 2006. Introducing elliptic, an R package for elliptic and modular functions. Journal of Statistical Software, 15(7).
}
