% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SDWBA_curved}
\alias{SDWBA_curved}
\title{Calculates the theoretical TS of a fluid-like scatterer at a given frequency
using the stochastic distorted Born wave approximation (DWBA) model using
Eq. (6) from Stanton et al. (1998).}
\usage{
SDWBA_curved(object)
}
\arguments{
\item{object}{FLS-class scatterer.}
}
\description{
Calculates the theoretical TS of a fluid-like scatterer at a given frequency
using the stochastic distorted Born wave approximation (DWBA) model using
Eq. (6) from Stanton et al. (1998).
}
\references{
Stanton, T.K., Chu, D., and Wiebe, P.H. 1998. Sound scattering by several
zooplankton groups. II. Scattering models. J. Acoust. Soc. Am., 103, 236-253.
}
