% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{sdwba_curved_initialize}
\alias{sdwba_curved_initialize}
\title{Initialize FLS-class object for SDWBA modeling}
\usage{
sdwba_curved_initialize(
  object,
  frequency,
  sound_speed_sw = 1500,
  density_sw = 1026,
  n_iterations = 100,
  n_segments_init = 14,
  phase_sd_init = base::sqrt(2)/2,
  length_init = 0.03835,
  frequency_init = 120000
)
}
\arguments{
\item{object}{FLS-class object.}

\item{frequency}{Frequency (Hz).}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}
}
\description{
Initialize FLS-class object for SDWBA modeling
}
