% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sdwba_resample}
\alias{sdwba_resample}
\title{Resample shape for SDWBA model with piecewise constant radius}
\usage{
sdwba_resample(object, n_segments)
}
\arguments{
\item{object}{FLS-class object to resample}

\item{n_segments}{Number of segments in the resampled shape}
}
\value{
FLS object with resampled shape
}
\description{
This function resamples the shape of a fluid-like scatterer (FLS) object for
use in stochastic distorted wave Born approximation (SDWBA) calculations.
The resampling preserves the overall shape of the scatterer while creating
a new representation with the specified number of segments. The radius
assignment uses a stepwise algorithm to maintain piecewise constant radius
values across segments.
}
