% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{segmentize}
\alias{segmentize}
\title{Discretize vector into separate intervals of different length}
\usage{
segmentize(x1, x0)
}
\arguments{
\item{x1}{Desired vector/interval}

\item{x0}{Original or initial vector/interval}
}
\description{
Discretize vector into separate intervals of different length
}
