% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{Shape}
\alias{Shape}
\alias{Shape-class}
\title{Generic scattering shape object used throughout this package.}
\value{
An object from the \code{Shape} S4 class hierarchy.
}
\description{
A S4 class that provides slots to contain relevant shape data and metadata
for a variety of arbitrary and canonical shapes and geometries. See
\link{Scatterer} for a more detailed description on how
this S4 object interacts with generic \link{Scatterer}
objects.
}
\section{Slots}{

\describe{
\item{\code{position_matrix}}{Position matrix that provides the 2D representation of
the body shape}

\item{\code{shape_parameters}}{A list of additional shape specifications}
}}

\examples{
methods::getClass("Shape")

}
\keyword{internal}
\keyword{shapes}
