% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{shape-class}
\alias{shape-class}
\title{Generic scattering shape object used throughout this package.}
\description{
A S4 class that provides slots to contain relevant shape data and metadata for
a variety of arbitrary and canonical shapes and geoemtries. See
\link[acousticTS]{scatterer-class} for a more detailed description on how
this S4 object interacts with generic \link[acousticTS]{scatterer-class}
objects.
}
