% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Shapely}
\alias{Shapely}
\title{Manipulate scatterer object.}
\usage{
Shapely(shape, curve, pc, theta, length)
}
\arguments{
\item{shape}{Input scatterer shape. [Currently only supports the FLS object-class]}

\item{curve}{Curve (boolean, T/F).}

\item{pc}{Radius of curvature (pc)}

\item{theta}{Orientation}

\item{length}{New Length (m)}
}
\value{
Returns manipulated object.
}
\description{
Manipulate scatterer object.
}
