% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,Scatterer-method}
\alias{show,Scatterer-method}
\title{Generic function for show(...) for different scatterers.}
\usage{
\S4method{show}{Scatterer}(object)
}
\arguments{
\item{object}{Scattering object.}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
Generic function for show(...) for different scatterers.
}
\examples{
show(cal_generate(material = "WC", n_segments = 20))
}
\keyword{internal}
