% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{sigma_bs}
\alias{sigma_bs}
\title{Calculates the linear backscattering coefficient (sigma_bs)}
\usage{
sigma_bs( f_bs )
}
\arguments{
\item{f_bs}{Linear scattering length (m), or related expression}
}
\value{
Returns the linear backscattering coefficient that can be converted to TS
}
\description{
Calculates the linear backscattering coefficient (sigma_bs)
}
