% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{Sphere}
\alias{Sphere}
\alias{Sphere-class}
\title{Spherical body shape}
\value{
An object of the \code{Sphere} S4 class.
}
\description{
Spherical body shape
}
\examples{
methods::getClass("Sphere")

}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
