% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{sphere.cw}
\alias{sphere.cw}
\title{Calculate the acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere of a certain material at a given frequency.}
\usage{
sphere.cw(frequency, c, rho, material = "Tungsten carbide",
  diameter = 38.1, pulse_length)
}
\arguments{
\item{frequency}{The acoustic frequency (Hz)}

\item{c}{The ambient seawater sound speed (m/s).}

\item{rho}{The ambient seawater density (kg/m^3).}

\item{material}{The material of the calibration sphere (tungsten carbide, aluminum, stainless steel, copper)}

\item{diameter}{The diameter of the sphere (mm).}

\item{pulse_length}{The pulse length (microseconds) of a signal.}
}
\description{
This function is a wrapper around TS_calculate(...) that parametrizes the remainder of the model, while also doing 
simple calculations that do not need to be looped. This function provides a mean TS which includes a FM sweep around
a center frequency. This is calculated by adding and subtracting half an estimated bandwidth (1/pulse length) and taking
the mean (calculated in the linear domain).
}
\examples{
frequency <- 120e3
c <- 1500
rho <- 1030
material <- "Tungsten carbide"
diameter <- 38.1
pulse_length <- 256
sphere.cw(frequency,c,rho,material,diameter,pulse_length)
[1] -39.549
#' @return 
The theoretical acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere at a given frequency and pulse length. 
}
