% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{sphere.spec}
\alias{sphere.spec}
\title{Calculate the frequency spectrum of the acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere .}
\usage{
sphere.spec(c, rho, material = "Tungsten carbide", diameter, fs, fe, fi)
}
\arguments{
\item{c}{The ambient seawater sound speed (m/s).}

\item{rho}{The ambient seawater density (kg/m^3).}

\item{material}{The material of the calibration sphere (tungsten carbide, aluminum, stainless steel, copper)}

\item{diameter}{The diameter of the sphere (mm).}

\item{fs}{Frequency start (Hz).}

\item{fe}{Frequency end (Hz).}

\item{fi}{Frequency interval (Hz).}
}
\value{
Returns the TS-frequency spectrum of an elastic sphere.
}
\description{
This function is a wrapper around TS_calculate(...) that parametrizes the remainder of the model, while also doing
simple calculations that do not need to be looped. This function provides TS estimates along a frequency spectrum.
}
\examples{
frequency <- 120e3
c <- 1500
rho <- 1030
material <- "Tungsten carbide"
diameter <- 38.1

sphere.spec(frequency,c,rho,material,diameter,fs=18e3,fe=230e3,fi=1e3)
}
