% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{sphere.spec_plot}
\alias{sphere.spec_plot}
\title{Plots the TS-frequency spectrum.}
\usage{
sphere.spec_plot(df)
}
\arguments{
\item{df}{A dataframe generated from the sphere.spec(...) function with two columns: Frequency and TS.}
}
\value{
Generates TS-frequency spectrum for an elastic sphere.
}
\description{
This function generates a plot of the TS-frequency spectrum using ggplot2.
}
\examples{
ts.df <- sphere.spec(frequency=120e3,c=1500,rho=1030,material="Tungsten carbide",diameter=38.1,fs=18e3,fe=230e3,fi=1e3)
sphere.spec_plot(ts.df)

}
