% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{sphere.ts}
\alias{sphere.ts}
\title{Calculate the acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere of a certain material at a given frequency.}
\usage{
sphere.ts(frequency, c, rho, material = "Tungsten carbide", diameter = 38.1)
}
\arguments{
\item{frequency}{The acoustic frequency (Hz)}

\item{c}{The ambient seawater sound speed (m/s).}

\item{rho}{The ambient seawater density (kg/m^3).}

\item{material}{The material of the calibration sphere (tungsten carbide, aluminum, stainless steel, copper)}

\item{diameter}{The diameter of the sphere (mm).}
}
\value{
The theoretical acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere at a given frequency.
}
\description{
This function is a wrapper around TS_calculate(...) that parametrizes the remainder of the model, while also doing
simple calculations that do not need to be looped. This function provides a TS estimate at a given frequency.
}
\examples{
frequency <- 120e3
c <- 1500
rho <- 1030
material <- "Tungsten carbide"
diameter <- 38.1
sphere.ts(frequency,c,rho,material,diameter)
[1] -39.52
}
