% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{sphms_initialize}
\alias{sphms_initialize}
\title{Initialize object for the modal series solution for a sphere}
\usage{
sphms_initialize(
  object,
  frequency,
  boundary = NULL,
  sound_speed_sw = 1500,
  density_sw = 1026,
  m_limit = NULL
)
}
\description{
Initialize object for the modal series solution for a sphere
}
