% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ssphere}
\alias{ssphere}
\title{Calculates theoretical TS of a solid sphere of a certain material at a given 
frequency.}
\usage{
ssphere(shape, frequency, c, rho, ...)
}
\arguments{
\item{shape}{CAL-class object.}

\item{frequency}{The acoustic frequency (Hz)}

\item{c}{The ambient seawater sound speed (m/s).}

\item{rho}{The ambient seawater density (kg/m^3).}

\item{a}{Optional. Radius (m).}

\item{c1}{Optional. Longitudinal sound speed (m/s).}

\item{c2}{Optional. Transversal sound speed (m/s).}

\item{rho1}{Optional. Density of sphere (kg/m^3).}

\item{g}{Optional. Density contrast.}

\item{h1}{Optional. Longitudinal sound speed contrast.}

\item{h2}{Optional. Transversal sound speed contrast.}
}
\value{
The theoretical acoustic target strength (TS, dB re: 1 m^2) of a 
solid sphere at a given frequency.
}
\description{
This function is a wrapper around TS_calculate(...) that parametrizes the 
remainder of the model, while also doing simple calculations that do not 
need to be looped. This function provides a TS estimate at a given frequency.
}
\references{
MacLennan D. N. (1981). The theory of solid spheres as sonar calibration 
targets. Scottish Fisheries Research No. 22, Department of Agriculture and 
Fisheries for Scotland.
}
