% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stanton_high_pass_initialize}
\alias{stanton_high_pass_initialize}
\title{Initialize object for Stanton high-pass approximation}
\usage{
stanton_high_pass_initialize(
  object,
  frequency,
  radius_shell = NULL,
  g_shell = NULL,
  h_shell = NULL,
  sound_speed_sw = 1500,
  density_sw = 1026
)
}
\arguments{
\item{object}{ESS-class object.}

\item{frequency}{Frequency (Hz).}

\item{radius_shell}{Radius of shell.}

\item{g_shell}{Optional shell density contrast.}

\item{h_shell}{Optional shell sound speed contrast.}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}
}
\description{
Initialize object for Stanton high-pass approximation
}
