% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{transmission_coefficient}
\alias{transmission_coefficient}
\title{Plane wave/plane interface transmission coefficient}
\usage{
transmission_coefficient(interface1, interface2, mode = "DWBA")
}
\arguments{
\item{interface1}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (1)}

\item{interface2}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (2)}

\item{mode}{Two options: coefficient calculation for "DWBA" and "KRM"}
}
\value{
Pressure-amplitude transmission coefficient at normal incidence.
}
\description{
Plane wave/plane interface transmission coefficient
}
\examples{
seawater <- data.frame(density = 1026, sound_speed = 1480)
animal <- data.frame(density = 1050, sound_speed = 1530)
transmission_coefficient(seawater, animal)
}
