% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{TS_calculate}
\alias{TS_calculate}
\title{Form function for calculating the acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere.}
\usage{
TS_calculate(q, q1, q2, alpha, beta, a)
}
\arguments{
\item{q}{The acoustic wavenumber (k) multiplied by the sphere radius (a) (Equation 6a).}

\item{q1}{The relationship between q (or ka) and the ratio of seawater-to-transversal soundspeed (Equation 6a).}

\item{q2}{The relationship between q (or ka) and the ratio of seawater-to-longitudinal soundspeed (Equation 6a).}

\item{alpha}{The relationship between the seawater-to-sphere soundspeed and density ratios (Equation 6d).}

\item{beta}{The relationship with alpha subtracted from between the seawater-to-sphere soundspeed 
and density ratios (Equation 6e).}

\item{a}{The radius of the target sphere (mm).}
}
\value{
The theoretical acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere.
}
\description{
This function calculates the acoustic target strength (TS, dB re: 1 m^2) of an elastic sphere. All equations are taken from 
MacLennan (1981).
}
