% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{vecnorm}
\alias{vecnorm}
\title{Vectorized Euclidean norm function}
\usage{
vecnorm(x)
}
\arguments{
\item{x}{A matrix with numeric, real values.}
}
\value{
Calculates the Euclidean norm of a vector.
}
\description{
Vectorized Euclidean norm function
}
\examples{
values <- matrix(c(1,2,3), ncol=3)
vecnorm(values)
# 3.741657
}
