% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{yd}
\alias{yd}
\title{Calculate the first derivative of the spherical Bessel function of the second kind.}
\usage{
yd(l,n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
yd(l,n) calculates the first derivative of the spherical Bessel function of the second kind (zbesy AMOS routine, Yv).
}
\description{
These functions alculate the first and second derivatives of the spherical Bessel functions required for the elastic sphere
TS model. These comprise the first derivatives of the spherical Bessel functions of the first and second kind,
as well as the second derivative of the spherical Bessel function of the first kind.
Functions are referenced from Amost D.E., "AMOS, A Portable Package for Bessel Functions of a Complex Argument
and Nonnegative Order", http://netlib.org/amos
}
\examples{
l <- 1
n < 1
yd(l,n)
[1] 2.223244
}
