% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{along_sum}
\alias{along_sum}
\title{Sums along the position vector}
\usage{
along_sum(rpos, iterations)
}
\arguments{
\item{rpos}{Position vector}

\item{iterations}{Number of iterations}
}
\description{
Sums along the position vector
}
