% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{along_sum}
\alias{along_sum}
\title{Along-matrix summing function}
\usage{
along_sum(rpos, iterations)
}
\arguments{
\item{rpos}{Position vector}

\item{iterations}{Number of iterations}
}
\description{
Along-matrix summing function
}
\keyword{internal}
