% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{Arbitrary-class}
\alias{Arbitrary-class}
\alias{Arbitrary}
\title{Arbitrary body shape}
\description{
Arbitrary body shape
}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
