% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{arbitrary}
\alias{arbitrary}
\title{Creates arbitrary body shape from user inputs}
\usage{
arbitrary(x_body, y_body, z_body, radius_body, length_units = "m")
}
\arguments{
\item{x_body}{x-axis (m)}

\item{y_body}{y-axis (m)}

\item{z_body}{z-axis (m)}

\item{radius_body}{Radius (m)}

\item{length_units}{Units for body length. Defaults to meters: "m"}
}
\description{
Creates arbitrary body shape from user inputs
}
\seealso{
\code{\link{Arbitrary}}
}
\keyword{shape_generation}
