% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{bbf_show}
\alias{bbf_show}
\title{show(...) for BBF-class objects.}
\usage{
bbf_show(object)
}
\arguments{
\item{object}{BBF-class object.}
}
\description{
show(...) for BBF-class objects.
}
\keyword{internal}
