% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark_ts}
\alias{benchmark_ts}
\title{Benchmark model outputs from Jech et al. (2015)}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(benchmark_ts)
}
\description{
Benchmark model outputs from Jech et al. (2015)
}
\keyword{datasets}
