% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-geometry.R
\name{brake}
\alias{brake}
\title{Support function for bending scatterer body shape and position matrix}
\usage{
brake(object, radius_curvature, mode = "ratio")
}
\arguments{
\item{object}{Dataframe or scatterer-class object}

\item{radius_curvature}{Radius of curvature that can be parameterized either
as a ratio relative to body length or actual measurement}

\item{mode}{Either "ratio" or "measurement"}
}
\description{
Support function for bending scatterer body shape and position matrix
}
\keyword{manipulator}
\keyword{shape}
