% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{bulk}
\alias{bulk}
\title{Calculate the bulk modulus (K).}
\usage{
bulk(E = NULL, G = NULL, nu = NULL)
}
\arguments{
\item{E}{Young's modulus (Pa).}

\item{G}{Shear modulus (Pa).}

\item{nu}{Poisson's ratio (Dimensionless).}
}
\value{
Bulk modulus (K, Pa).
}
\description{
Calculates the bulk modulus (K) from two of the three other elastic moduli:
Young's modulus (E), shear modulus (G), or Poisson's ratio (\eqn{\nu}).
Assumes 3D material properties.

The relationships used are:
\deqn{K = \frac{E G}{3(3G - E)}}
\deqn{K = \frac{2G(1 + \nu)}{3(1 - 2\nu)}}
\deqn{K = \frac{E}{3(1 - 2\nu)}}
}
\keyword{elastic}
