% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{bulk}
\alias{bulk}
\title{Calculate the bulk modulus (K).}
\usage{
bulk(E, G, nu)
}
\arguments{
\item{E}{Young's modulus (GPa).}

\item{G}{Shear modulus (GPa).}

\item{nu}{Poisson's ratio (unitless).}
}
\value{
Returns an estimate for the bulk modulus (K).
}
\description{
Calculate the bulk modulus (K) from two of the three other elastic
moduli to calculate the Lamé's parameter. When more than two values are input,
the function will default to using Young's (E) and shear (G) moduli. This
assumes that the input values represent 3D material properties.
}
