% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{bulk}
\alias{bulk}
\title{Calculate the bulk modulus (K).}
\usage{
bulk(E = NULL, G = NULL, nu = NULL)
}
\arguments{
\item{E}{Young's modulus (Pa).}

\item{G}{Shear modulus (Pa).}

\item{nu}{Poisson's ratio (Dimensionless).}
}
\value{
Returns an estimate for the bulk modulus (K).
}
\description{
Calculate the bulk modulus (K) from two of the three other elastic
moduli to calculate the Lamés parameter. When more than two
values are input, the function will default to using Young's (E) and shear
(G) moduli. This assumes that the input values represent 3D material
properties.
}
\keyword{elastic}
