% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CAL-class}
\alias{CAL-class}
\alias{CAL}
\title{Solid and calibration sphere (CAL) object/class.}
\arguments{
\item{material}{Material-type for the soldi sphere. See 'Details' built-in
material options.}

\item{a}{Spherical radius (m).}

\item{c1}{Longitudinal sound speed (m/s).}

\item{c2}{Transversal sound speed (m/s).}

\item{rho1}{Density (kg/m^3)}
}
\description{
A S4 class that provides slots to contain relevant metadata for solid sphere
objects belonging to the CAL-class scatterers.
}
\details{
There are several options for the \strong{material} argument:
\tabular{rlllll}{
 \strong{Material} \tab \strong{Argumment} \tab \strong{c1} \tab \strong{c2} 
 \tab \strong{\eqn{\rho1}}\cr
 \emph{Tungsten carbide} \tab "WC" \tab 6853 \tab 4171 \tab 8360\cr
 \emph{Stainless steel} \tab "steel" \tab 5980 \tab 3297 \tab 7970\cr
 \emph{Brass} \tab "brass" \tab 4372 \tab 2100 \tab 8360\cr
 \emph{Copper} \tab "Cu" \tab 4760 \tab 2288.5 \tab 8947\cr
 \emph{Aluminum} \tab "Al" \tab 6260 \tab 3080 \tab 2700\cr
}
}
