% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{CAL-class}
\alias{CAL-class}
\alias{CAL}
\title{Solid and calibration sphere (CAL) object/class.}
\description{
A S4 class that provides slots to contain relevant metadata for solid sphere
objects belonging to the CAL-class scatterers. This object is created using
parameters specific to the outer shell. The default behavior of this object is
to only reference these outer elastic shell properties with few exceptions that
are model-dependent. See \link{Scatterer} for a more
detailed description on how this S4 object is organized.
}
\seealso{
\link{Scatterer}
}
\keyword{scatterer_type}
