% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{cal_generate}
\alias{cal_generate}
\title{Generate a CAL-class object.}
\usage{
cal_generate(
  material = "WC",
  diameter = 0.0381,
  sound_speed_longitudinal = NULL,
  sound_speed_transversal = NULL,
  density_sphere = NULL,
  ID = NULL,
  diameter_units = "m"
)
}
\arguments{
\item{material}{Material-type for a solid sphere. See \code{Details} for available
options. Default is tungsten carbide (WC).}

\item{diameter}{Spherical diameter (m).}

\item{sound_speed_longitudinal}{Longitudinal sound speed (m/s).}

\item{sound_speed_transversal}{Transversal sound speed (m/s).}

\item{density_sphere}{Density (kg/m^3).}

\item{ID}{Optional metadata ID input.}

\item{diameter_units}{Units for diameter. Defaults to "m".}
}
\value{
Generates a CAL-class object.
}
\description{
Generate a CAL-class object.
}
