% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod}
\alias{cod}
\title{Sample code shape with fully inflated swimbladder.}
\format{
A pre-generated SBF scatterer containing all information required for
target strength modeling.
\describe{
\item{metadata}{Relevant and identifying metadata (\code{list}).}
\item{model_parameters}{Container for specified model parameters
(\code{list}).}
\item{model}{Model outputs and results (\code{list}).}
\item{body}{List with:
\itemize{
\item \code{rpos}: Position matrix (x, yw, zU, zL; m).
\item \code{sound_speed}: Flesh sound speed (\eqn{c_{body}}, m/s).
\item \code{density}: Flesh density (\eqn{\rho_{body}}, kg/m\eqn{^3}).
\item \code{theta}: Orientation relative to transducer
(\eqn{\theta_{body}}, radians).
}
}
\item{bladder}{List with:
\itemize{
\item \code{rpos}: Position matrix (x, yw, zU, zL; m).
\item \code{sound_speed}: Flesh sound speed (\eqn{c_{bladder}}, m/s).
\item \code{density}: Flesh density
(\eqn{\rho_{bladder}}, kg/m\eqn{^3}).
\item \code{theta}: Orientation relative to transducer
(\eqn{\theta_{bladder}}, radians).
}
}
\item{shape_parameters}{Named list with:
\itemize{
\item \code{body}: List with length (m), ncyl (int),
theta_units (str), length_units (str).
\item \code{bladder}: List with length (m), ncyl (int),
theta_units (str), length_units (str).
}
}
}
}
\usage{
data(cod)
}
\description{
Sample code shape with fully inflated swimbladder.
}
\keyword{datasets}
