% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod}
\alias{cod}
\title{Sample sardine shape with fully inflated swimbladder.}
\format{
A pre-generated SBF scatterer containing all information required for
target strength modeling.
\describe{
\item{metadata}{Relevant and identifying metadata (\code{list}).}
\item{model_parameters}{Container for specified model parameters (\code{list}).}
\item{model}{Model outputs and results (\code{list}).}
\item{body}{
\itemize{
\item{rpos:}{ Position matrix (x, yw, zU, zL; m).}
\item{sound_speed:}{ Flesh sound speed (\ifelse{html}{\out{c<sub>body</sub>}}{\eqn{c_{body}}}, m \ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}})}
\item{density:}{ Flesh density (\ifelse{html}{\out{&rho;<sub>body</sub>}}{\eqn{\rho_{body}}}, kg \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}})}
\item{theta:}{ Body orientation relative to transmitting transducer (\ifelse{html}{\out{&theta;<sub>body</sub>}}{\eqn{\theta_{body}}}
, rads)}}}
\item{bladder}{
\itemize{
\item{rpos:}{ Position matrix (x, yw, zU, zL; m).}
\item{sound_speed:}{ Flesh sound speed (\ifelse{html}{\out{c<sub>bladder</sub>}}{\eqn{c_{bladder}}}, m \ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}})}
\item{density:}{ Flesh density (\ifelse{html}{\out{&rho;<sub>bladder</sub>}}{\eqn{\rho_{bladder}}}, kg \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}})}
\item{theta:}{ Body orientation relative to transmitting transducer (\ifelse{html}{\out{&theta;<sub>bladder</sub>}}{\eqn{\theta_{bladder}}}
, rads)}}}
\item{shape_parameters}{
\itemize{
\item{body}{
\itemize{
\item{length:}{ Body length (m).}
\item{ncyl:}{ Number of discrete cylinders along body length.}
\item{theta_units:}{ Units used for orientation of body.}
\item{length_units:}{ Units used for body length.}}}
\item{bladder}{
\itemize{
\item{length:}{ Bladder length (m).}
\item{ncyl:}{ Number of discrete cylinders along swimbladder length.}
\item{theta_units:}{ Units used for orientation of swimbladder.}
\item{length_units:}{ Units used for swimbladder length.
}}}}}}
}
\usage{
data(cod)
}
\description{
Sample sardine shape with fully inflated swimbladder.
}
\keyword{datasets}
