% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{compressibility}
\alias{compressibility}
\title{Calculate she compressibility (\eqn{\kappa}) of a scattering
boundary/interface.}
\usage{
compressibility(medium, target)
}
\arguments{
\item{medium}{Dataframe object containing density (\eqn{kgm^{-3}}) and
sound speed (\eqn{ms^{-1}}) values for a fluid medium external to a
scattering interface (e.g., seawater).}

\item{target}{Dataframe object containing density (\eqn{kgm^{-3}}) and
sound speed (\eqn{ms^{-1}}) values for a target boundary.}
}
\value{
Compressibility contrast (\eqn{\kappa}), dimensionless.
}
\description{
Calculates the compressibility contrast (\eqn{\kappa}) between a scattering
interface and the surrounding medium. Compressibility is defined as:
\deqn{
 K = \frac{1}{\rho c^2}
}
where \eqn{\rho} is density (\eqn{kg~m^{-3}}) and \eqn{c} is sound speed
(\eqn{m~s^{-1}}).

The compressibility contrast is then:
\deqn{
 \kappa = \frac{K_2 - K_1}{K_1}
}

where \eqn{K_1} is the compressibility of the medium and \eqn{K_2} is that
of the target interface.
}
\examples{
seawater <- data.frame(density = 1026, sound_speed = 1480)
animal <- data.frame(density = 1050, sound_speed = 1530)
compressibility(seawater, animal)
}
