% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{contour_integrate}
\alias{contour_integrate}
\title{Complex integration}
\usage{
contour_integrate(integral, x, y)
}
\arguments{
\item{integral}{Integral function used for numerical integration via adaptive
quadrature}

\item{x}{Indexing argument for multi-row objects}

\item{y}{Indexing argument for multi-column objects}
}
\description{
Complex integration
}
