% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{create_shape}
\alias{create_shape}
\title{A wrapper function that automatically creates generalized and/or canonical
shapes for TS modeling.}
\usage{
create_shape(shape, ...)
}
\arguments{
\item{shape}{Shape. Details for shape specification are provided under
'Details', including mandatory arguments.}

\item{...}{Additional input arguments for subsequent shape generation
functions.}
}
\value{
Chu, D., Foote, K.G., and Stanton, T.K. 1993. Further analysis of target
strength measurements of Antarctic krill at 38 and 120 kHz: Comparison and
deformed cylinder model and inference of orientation distribution. The
Journal of the Acoustical Society of America, 93(5): 2985-2988.
https://doi.org/10.1121/1.405818

Smith, J.N., Ressler, P.H., and Warren, J.D. 2013. A distorted wave Born
approximation target strength model for Bering Sea euphausiids. ICES Journal
of Marine Science, 70(1): 204-214. https://doi.org/10.1093/icesjms/fss140
}
\description{
A wrapper function that automatically creates generalized and/or canonical
shapes for TS modeling.
}
\details{
The \strong{shape} argument specifies what shape for the function to
generate the desired shape for TS modeling. Options currently include:
\tabular{rlllll}{
\tab \strong{Object shape} \tab \strong{shape = ...} \tab  \tab
\strong{Parameters} \tab \strong{Root function}\cr
\tab \emph{Discrete/tapered cylinder} \tab "cylinder" \tab
\tab length, radius \tab
\code{\link[=cylinder]{cylinder(...)}}\cr
\tab \emph{Polynomial cylinder} \tab "polynomial_cylinder" \tab
\tab length, radius, polynomial \tab
\code{\link[=polynomial_cylinder]{polynomial_cylinder(...)}}\cr
\tab \emph{Prolate spheroid} \tab "prolate_spheroid" \tab
\tab length, radius \tab
\code{\link[=prolate_spheroid]{prolate_spheroid(...)}}\cr
\tab \emph{Sphere} \tab "sphere" \tab \tab radius \tab
\code{\link[=sphere]{sphere(...)}}\cr
}

\subsection{Model Parameter Definitions}{
\itemize{
\item \strong{length}: the x-axis length of the shape.
\item \strong{radius}: the radius of the shape when applicable.
\item \strong{length_radius_ratio}: the length-to-radius ratio (L/A), which
specifically refers to the radius at the mid-point of the cylinder and
should be the maximum value. A typical L/A ratio in the literature is 16
for krill.
\item \strong{taper}: the taper order (n), which parameterizes the tapering
function reported by Chu \emph{et al.} (1993) to create a tapered cylinder.
The tapering order will converge on a prolate and oblate spheroid when
L > 2a and L < 2a, respectively, and n = 2. A typical taper order in the
literature is 10.
\item \strong{polynomial}: the vector of arbitrary polynomial coefficients
to generate a deformed cylinder as reported by Smith \emph{et al.} (2013).
Although listed as a mandatory argument for the polynomial cylinder
function, it has a default setting that uses the sixth-degree polynomial
coefficients reported by Smith \emph{et al.} (2013).
}
}
}
\keyword{shape_generation}
