% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{cylinder}
\alias{cylinder}
\title{Creates a cylinder.}
\usage{
cylinder(length, radius, length_radius_ratio, taper, n_segments)
}
\arguments{
\item{length}{Length (m).}

\item{radius}{Maximum/uniform radius (m).}

\item{length_radius_ratio}{Optional ratio input when radius is not explicitly
known.}

\item{taper}{Optional input that is the degree of taper to round ends of
the cylinder.}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
1e2 segments.}
}
\value{
Creates the position vector for a tapered or untapered cylinder.
}
\description{
Creates a cylinder.
}
