% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{cylinder}
\alias{cylinder}
\title{Creates a cylinder.}
\usage{
cylinder(length_body, radius_body, length_radius_ratio, 
taper, n_segments, length_units)
}
\arguments{
\item{length_body}{Length (m).}

\item{radius_body}{Maximum/uniform radius (m).}

\item{length_radius_ratio}{Optional ratio input when radius is not explicitly
known.}

\item{taper}{Optional input that is the degree of taper to round ends of
the cylinder.}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
1e2 segments.}

\item{length_units}{Units (default is meters, "m").}
}
\value{
Creates the position vector for a tapered or untapered cylinder.
}
\description{
Creates a cylinder.
}
