% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{DCM}
\alias{DCM}
\title{Calculates the theoretical TS of a target using the deformed cylinder
model (DCM).}
\usage{
DCM(object)
}
\arguments{
\item{object}{FLS-class scatterer.}
}
\value{
Target strength (TS, dB re. 1 \eqn{m^2}) of a FLS-class object.
#' @references
Stanton, T.K., Chu, D., and Wiebe, P.H. 1998. Sound scattering by several
zooplankton groups. II. Scattering models. Journal of the Acoustical Society
of America, 103(1), 236-253.
}
\description{
Calculates the theoretical TS of a target using the deformed cylinder
model (DCM).
}
