% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{dcm_initialize}
\alias{dcm_initialize}
\title{Initialize object for modeling using the DCM.}
\usage{
dcm_initialize(
  object,
  frequency,
  radius_cylinder = NULL,
  radius_curvature = NULL,
  radius_curvature_ratio = 3,
  radius_cylinder_fun = "max",
  length = NULL,
  g = NULL,
  h = NULL,
  theta = NULL,
  sound_speed_sw = 1500,
  density_sw = 1026,
  alpha_B = 0.8
)
}
\arguments{
\item{object}{FLS-class object.}

\item{frequency}{Transmit frequency (kHz)}

\item{radius_cylinder}{Optional input to override current shape radius.}

\item{radius_curvature}{Numeric input for the radius of curvature}

\item{radius_curvature_ratio}{Ratio between body length and the radius of
curvature. Defaults to 3.0.}

\item{radius_cylinder_fun}{Defines which radius value will be used from the
radius vector. Defaults to "max", but also accepts "mean" and "median".}

\item{length}{Body length (m).}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Body orientation relative to the incident sound wave.}

\item{sound_speed_sw}{Seawater sound speed
(\ifelse{html}{\out{c<sub>body</sub>}}{\eqn{c_{body}}},
m \ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}}).}

\item{density_sw}{Seawater density
(\ifelse{html}{\out{&rho;<sub>body</sub>}}{\eqn{\rho_{body}}},
kg \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}})}

\item{alpha_B}{Numerical coefficient
(\ifelse{html}{\out{&alpha;<sub>B</sub>}}{\eqn{\alpha_{B}}}).}
}
\description{
Initialize object for modeling using the DCM.
}
