% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{dwba_initialize}
\alias{dwba_initialize}
\title{Initialize FLS-class object for TS modeling.}
\usage{
dwba_initialize(
  object,
  frequency,
  sound_speed_sw = 1500,
  density_sw = 1026,
  theta = pi/2
)
}
\arguments{
\item{object}{FLS-class object.}

\item{frequency}{Frequency (Hz).}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}

\item{theta}{Angle of incident soundwave (pi / 2 is broadside).}
}
\description{
Initialize FLS-class object for TS modeling.
}
