% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{ESS-class}
\alias{ESS-class}
\alias{ESS}
\title{Elastic shelled scatterer (ESS) object/class.}
\description{
A S4 class that provides slots to contain relevant metadata for elastic
shelled scatterers/objects belonging to the ESS-class scatterers.
}
