% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ESS-class}
\alias{ESS-class}
\alias{ESS}
\title{Elastic shelled scatterer (ESS) object/class.}
\arguments{
\item{a}{Radius (m).}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Orientation (\eqn{\theta}, radians)}

\item{L}{Optional. Length of the longitudinal axis (m) for non-spherical 
shapes.}

\item{spherical}{Optional. Boolean for specifying whether scatterer is 
spherical or oblong.}
}
\description{
A S4 class that provides slots to contain relevant metadata for elastic 
shelled scatterers/objects belonging to the ESS-class scatterers.
}
