% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{ess_generate}
\alias{ess_generate}
\title{Generate ESS shape}
\usage{
ess_generate(
  shape = "sphere",
  x_body = NULL,
  y_body = NULL,
  z_body = NULL,
  radius_shell,
  shell_thickness = NULL,
  g_fluid = NULL,
  density_fluid = NULL,
  h_fluid = NULL,
  sound_speed_fluid = NULL,
  g_shell = NULL,
  density_shell = NULL,
  h_shell = NULL,
  sound_speed_shell = NULL,
  E = NULL,
  G = NULL,
  K = NULL,
  nu = NULL,
  theta_shell = pi/2,
  ID = NULL,
  theta_units = "radians",
  length_units = "m"
)
}
\arguments{
\item{shape}{Optional input argument that dictates shape-type, if desired, for
generalized shapes.}

\item{x_body}{Vector containing x-axis body (m) shape data.}

\item{y_body}{Vector containing y-axis body (m) shape data.}

\item{z_body}{Vector containing z-axis body (m) shape data.}

\item{radius_shell}{Radius of shell (m).}

\item{shell_thickness}{Optional shell thickness (m).}

\item{g_fluid}{Optional density contrast for fluid-like body.}

\item{density_fluid}{Optional density for fluid-like body (kg/m³).}

\item{h_fluid}{Optional sound speed contrast for fluid-like body.}

\item{sound_speed_fluid}{Optional sound speed for fluid-like body (m/s).}

\item{g_shell}{Density contrast for the shell.}

\item{density_shell}{Optional density for the shell (kg/m³).}

\item{h_shell}{Sound speed contrast for the shell.}

\item{sound_speed_shell}{Optional sound speed for the shell (m/s).}

\item{E}{Young's modulus (Pa) of the shell material.}

\item{G}{Shear modulus (Pa) of the shell material.}

\item{K}{Bulk modulus (Pa) of the shell material.}

\item{nu}{Poisson's ratio (Dimensionless) of the shell material.}

\item{theta_shell}{Object orientation relative to incident sound wave.}

\item{ID}{Optional metadata entry.}

\item{theta_units}{Units used for orientation. Defaults to "radians".}

\item{length_units}{Units used for position vector. Defaults to "m".}
}
\value{
ESS-class object
}
\description{
Generate ESS shape
}
