% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{ess_show}
\alias{ess_show}
\title{show(...) for ESS-class objects.}
\usage{
ess_show(object)
}
\arguments{
\item{object}{ESS-class object.}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
show(...) for ESS-class objects.
}
\keyword{internal}
