% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract}
\alias{extract}
\title{Primary accessor function for dredging specific data from scatterer objects}
\usage{
extract(object, feature)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{feature}{Feature of interest (e.g. body).}
}
\description{
Primary accessor function for dredging specific data from scatterer objects
}
