% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{extract}
\alias{extract}
\title{Primary access function.}
\usage{
extract(object, feature)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{feature}{Feature of interest (e.g. body).}
}
\description{
Primary access function.
}
