% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FLS-class}
\alias{FLS-class}
\alias{FLS}
\title{Fluid-like scatterer (FLS) object/class.}
\arguments{
\item{rpos}{A 3xN matrix, where N is the number of columns, or segments, in our scatterer object. The rows represent the x-, y-, and z-axis
shape values in that order. Units in meters.}

\item{a}{An N-length radius vector. Units in meters.}

\item{g}{Density contrast. Units are dimensionless.}

\item{h}{Sound speed contrast. Units are dimensionless.}

\item{theta}{Animal orientation, where \eqn{\pi/2} is broadside incidence. Units in radians.}

\item{curve}{Body curvature/flexure. Boolean (T/F) which flags the appropriate model formulation.}

\item{pc}{Radius of curvature ratio. Units are dimensionless.}

\item{L}{Length of target. Units in meters.}

\item{ncyl}{Number of segments/cylinders comprising the shape. Units should be an integer value.}
}
\description{
A S4 class that provides slots to contain relevant animal metadata for parameterizing models for fluid-like scatterers (FLS) partitioned
into discretized cylinders. This, specifically, includes a position matrix, radius, material properties (g, h), orientation,
animal shape, and body curvature. This class is used within the DWBA and DFCM model functions. In the future, this will also allow for
converting one class of scatterer into another for seemless usage for model comparisons.
}
\details{
FLS objects can be created using \link[acousticTS]{FLSwrite} and \link[acousticTS]{FLSgenerate}. These objects can also be read using
built-in datasets (e.g., \link[acousticTS]{mcgehee}) or using \link[acousticTS]{FLSread}. The \link[acousticTS]{SDWBA}, \link[acousticTS]{SDWBA.sim},
\link[acousticTS]{DFCM}, and \link[acousticTS{Shapely}] functions all use FLS-objects for their respective inputs.
}
\seealso{
\code{\link[acousticTS{FLSread}}
}
