% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{fls_generate}
\alias{fls_generate}
\title{Manually generate a FLS object.}
\usage{
fls_generate(
  x_body,
  y_body,
  z_body,
  radius_body,
  g_body,
  h_body,
  theta_body = pi/2,
  theta_units = "radians",
  length_units = "m",
  ID = NULL
)
}
\arguments{
\item{x_body}{Vector containing x-axis body (m) shape data.}

\item{y_body}{Vector containing y-axis body (m) shape data.}

\item{z_body}{Vector containing z-axis body (m) shape data.}

\item{radius_body}{Vector containing radii (m).}

\item{g_body}{Density contrast.}

\item{h_body}{Soundspeed contrast}

\item{theta_body}{Orientation of the target relative to the transmit source
(\eqn{\theta}). Broadside incidence is considered 90 degrees, or pi/2.
Default value is pi/2; input should be in radians.}

\item{theta_units}{Units used for orientation. Defaults to "radians".}

\item{length_units}{Units used for position vector. Defaults to "m".}

\item{ID}{Optional metadata entry.}
}
\value{
Calls in an FLS-class object from a *.csv file
}
\description{
Manually generate a FLS object.
}
