% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{fls_generate}
\alias{fls_generate}
\title{Manually generate a FLS object.}
\usage{
fls_generate(
  shape = "arbitrary",
  x_body = NULL,
  y_body = NULL,
  z_body = NULL,
  length_body = NULL,
  radius_body = NULL,
  radius_curvature_ratio = NULL,
  n_segments = 18,
  g_body,
  h_body,
  theta_body = pi/2,
  ID = NULL,
  length_units = "m",
  theta_units = "radians",
  ...
)
}
\arguments{
\item{shape}{Optional input argument that dictates shape-type, if desired,
for generalized shapes.}

\item{x_body}{Vector containing x-axis body (m) shape data.}

\item{y_body}{Vector containing y-axis body (m) shape data.}

\item{z_body}{Vector containing z-axis body (m) shape data.}

\item{length_body}{Optional input for a generic length value input.}

\item{radius_body}{Vector containing radii (m).}

\item{radius_curvature_ratio}{Length-to-curvature ratio (pc/L).}

\item{n_segments}{Number of body segments.}

\item{g_body}{Density contrast. This can either be a single value (i.e.
homogenous) or a vector of values (i.e. inhomogenous).}

\item{h_body}{Soundspeed contrast. This can either be a single value (i.e.
homogenous) or a vector of values (i.e. inhomogenous).}

\item{theta_body}{Orientation of the target relative to the transmit source
(\eqn{\theta}). Broadside incidence is considered 90 degrees, or pi/2.
Default value is pi/2; input should be in radians.}

\item{ID}{Optional metadata entry.}

\item{length_units}{Units used for position vector. Defaults to "m".}

\item{theta_units}{Units used for orientation. Defaults to "radians".}

\item{...}{Additional parameters.}
}
\value{
FLS-class object
}
\description{
Manually generate a FLS object.
}
\seealso{
\code{\link{FLS}}
}
\keyword{scatterer_type_generation}
