% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{fls_show}
\alias{fls_show}
\title{show(...) for FLS-class objects.}
\usage{
fls_show(object)
}
\arguments{
\item{object}{FLS-class object.}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
show(...) for FLS-class objects.
}
\keyword{internal}
