% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{gas_generate}
\alias{gas_generate}
\title{Create GAS object}
\usage{
gas_generate(
  radius_body,
  g_body = 0.0012,
  h_body = 0.22,
  ID = NULL,
  diameter_units = "m"
)
}
\arguments{
\item{radius_body}{Vector containing radii (m).}

\item{g_body}{Density contrast.}

\item{h_body}{Soundspeed contrast}

\item{ID}{Optional metadata entry.}

\item{diameter_units}{Diameter units. Defaults to "m".}
}
\value{
Creates a FLS-class object from a *.csv file
}
\description{
Create GAS object
}
