% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{hc}
\alias{hc}
\alias{hcd}
\title{Cylindrical Bessel (Hankel) function of the third kind (Hc)}
\usage{
hc( l , n )

hcd( l , n )
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Bessel function of the second kind (H_v). Functions are
referenced from Amost D.E., "AMOS, A Portable Package
for Bessel Functions of a Complex Argument and Nonnegative Order",
http://netlib.org/amos
}
\description{
Cylindrical Bessel (Hankel) function of the third kind (Hc)

First derivative of the Bessel function of the third kind (Hcd)
}
