% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{hc}
\alias{hc}
\alias{hcd}
\title{Cylindrical Bessel (Hankel) function of the third kind and its derivatives}
\usage{
hc(l, n)

hcd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the cylindrical Bessel function of the first kind (\eqn{H_v}) and
its respective derivatives.
}
\description{
Cylindrical Bessel (Hankel) function of the third kind and its derivatives
}
\references{
Amos D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
\keyword{bessel3}
