% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{high_pass_stanton}
\alias{high_pass_stanton}
\title{Calculates the theoretical TS of a shelled organism using the non-modal
High Pass (HP) model}
\usage{
high_pass_stanton(object)
}
\arguments{
\item{object}{Desired animal object (Elastic Shelled).}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a shelled organism using the non-modal
High Pass (HP) model
}
\references{
Stanton, T.K. (1989). Simple approximate formulas for backscattering of
sound by spherical and elongated objects. The Journal of the Acoustical
Society of America, 86, 1499-1510.
}
