% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP.R
\name{HP}
\alias{HP}
\title{Calculates the theoretical TS of a shelled organism using the non-modal High Pass (HP) model}
\usage{
HP(animal, c = 1500, frequency, a = max(animal@a))
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a shelled organism using the non-modal High Pass (HP) model
}
