% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP.R
\name{HP}
\alias{HP}
\title{Calculates the theoretical TS of a shelled organism using the non-modal High Pass (HP) model}
\usage{
HP(
  shape = NULL,
  c = 1500,
  frequency,
  g = shape@g,
  h = shape@h,
  a = max(shape@a)
)
}
\arguments{
\item{shape}{Desired object/animal shape.}

\item{c}{Sound speed of surrounding medium (m/s). Default value is 1500 m/s.}

\item{frequency}{Frequency (Hz).}

\item{a}{Radius of object (m).}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a shelled organism using the non-modal High Pass (HP) model
}
\references{

}
