% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{HP}
\alias{HP}
\title{Calculates the theoretical TS of a shelled organism using the non-modal High Pass (HP) model}
\usage{
HP(shape = NULL, c = 1500, frequency, g = shape@g, h = shape@h, a = shape@a)
}
\arguments{
\item{shape}{Desired animal object. Must be class FS.}

\item{c}{Sound speed of surrounding medium (m/s). Default value is 1500 m/s.}

\item{frequency}{Frequency (Hz).}

\item{g}{Optional. Density contrast.}

\item{h}{Optional. Sound speed contrast.}

\item{a}{optional. Radius vector of sphere (m).}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS of a shelled organism using the non-modal High Pass (HP) model
}
\references{
Lavery, A.C., Wiebe, P.H., Stanton, T.K., Lawson, G.L., Benfield, M.C., 
Copley, N. 2007. Determining dominant scatterers of sound in mixed 
zooplankton popuilations. The Journal of the Acoustical Society of America,
122(6): 3304-3326.
}
