% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{hs}
\alias{hs}
\alias{hsd}
\title{Spherical Bessel function of the third kind (ha)}
\usage{
hs( l , n )

hsd( l , n )
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\description{
Spherical Bessel function of the third kind (ha)

First derivative of the spherical Bessel function of the third kind (had)
}
