% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{hs}
\alias{hs}
\alias{hsd}
\title{Spherical Bessel function of the third kind}
\usage{
hs(l, n)

hsd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\description{
Spherical Bessel function of the third kind

First derivative of the spherical Bessel function of the third kind
}
