% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{ja}
\alias{ja}
\title{Calculate Bessel function of the first kind.}
\usage{
ja(l,n)
#examples
l <- 1.5
n <- 1
ja(l,n)
[1] 0.2402978
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Bessel function of the first kind (J_a). Functions are referenced from Amost D.E., "AMOS, A Portable Package 
for Bessel Functions of a Complex Argument and Nonnegative Order", http://netlib.org/amos
}
\description{
Calculate Bessel function of the first kind.
}
