% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{ja}
\alias{ja}
\alias{jd}
\alias{jdd}
\title{Bessel functions of the first kind}
\usage{
ja(l,n)

jd(l, n)

jdd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Bessel function of the first kind (J_a).
}
\description{
Bessel functions of the first kind

First derivative of Bessel function of the first kind

Second derivative of the Bessel function of first kind
}
\examples{
l <- 1.5
n <- 1
ja(l,n)
# 0.2402978
}
\references{
Amost D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
