% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{jc}
\alias{jc}
\alias{jcd}
\alias{jcdd}
\title{Wrapper function for the Cylindrical Bessel function of the first kind (Jc)}
\usage{
jc( l , n )

jcd( l , n )

jcdd( l , n )
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Cylindrical Bessel function of the first kind (J_a).
}
\description{
Wrapper function for the Cylindrical Bessel function of the first kind (Jc)

First derivative of the Cylindrical Bessel function of the first kind (Jcd)

Second derivative of the Cylindrical Bessel function of the first kind (Jcdd)
}
\references{
Amost D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
