% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{jc}
\alias{jc}
\alias{jcd}
\alias{jcdd}
\title{Cylindrical Bessel function of the first kind and its respective derivatives}
\usage{
jc(l, n)

jcd(l, n)

jcdd(l, n)
}
\arguments{
\item{l}{A real integer or fractional order}

\item{n}{A real argument}
}
\value{
Calculates the cylindrical Bessel function of the first kind (\eqn{J_v}) and
its respective derivatives.
}
\description{
Calculate the cylindrical Bessel function of the first kind (jc) and both its
first (jcd) and second (jcdd) derivatives.
}
\references{
Amos D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
\keyword{bessel1}
