% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{jl}
\alias{jl}
\title{Calculate spherical Bessel function of the first kind.}
\usage{
jl(l,n,sign)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}

\item{sign}{Flag used to to determine whether to add or subtract 0.5 from the function's order (l); only used
for spherical Bessel function of first kind.}
}
\value{
jl(l,n,sign) calculates the spherical Bessel function of the first kind (zbesj AMOS routine, Iv).
}
\description{
This function calculates the spherical Bessel functions of the first kind. Referenced from Amost D.E.,
"AMOS, A Portable Package for Bessel Functions of a Complex Argument and Nonnegative Order", http://netlib.org/amos
}
\examples{
l <- 1
n < 1
jl(l,n,sign=1)
[1] 0.3011687
}
