% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{js}
\alias{js}
\alias{jsd}
\alias{jsdd}
\title{Spherical Bessel function of the first kind.}
\usage{
js(l,n)

jsd(l, n)

jsdd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
js(l,n) calculates the spherical Bessel function of the first kind 
(zbesj AMOS routine, Iv).
}
\description{
Spherical Bessel function of the first kind.

Calculate the first derivative of the spherical Bessel function of the 
first kind.

Calculate the second derivative of the spherical Bessel function of the 
first kind.
}
\examples{
l <- 1
n <- 1
js(l,n)
# 0.3011687
}
