% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{js}
\alias{js}
\alias{jsd}
\alias{jsdd}
\title{Spherical Bessel function of the first kind and its respective derivatives}
\usage{
js(l, n)

jsd(l, n)

jsdd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the spherical Bessel function of the first kind (js) and both its
first (jsd) and second (jsdd) derivatives.
}
\description{
Spherical Bessel function of the first kind and its respective derivatives
}
\references{
Amos D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
