% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kcalc}
\alias{kcalc}
\title{Calculate the acoustic wavenumber based on the sound speed of water.}
\usage{
kcalc(f,c)
}
\arguments{
\item{f}{Frequency (Hz)}

\item{c}{Sound speed (m/s)}
}
\value{
Calculates the acoustic wavenumber based on the sound speed of water
}
\description{
Calculate the acoustic wavenumber based on the sound speed of water.
}
\examples{
c <- 1500 #m/s
f <- 120e3 #Hz
kcalc(f,c)
[1] 502.6547
}
