% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{kcalc}
\alias{kcalc}
\title{Calculate the acoustic wavenumber (k) based on the sound speed of water.}
\usage{
kcalc( frequency , sound_speed )
}
\arguments{
\item{frequency}{Frequency (f, Hz)}

\item{sound_speed}{Sound speed (c, m/s)}
}
\value{
Calculates the acoustic wavenumber (k) based on the sound speed of water
}
\description{
Calculate the acoustic wavenumber (k) based on the sound speed of water.
}
