% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{KRM}
\alias{KRM}
\title{Calculates the theoretical TS using Kirchoff-ray Mode approximation.}
\usage{
KRM(object)
}
\arguments{
\item{object}{Desired object/animal shape. Must be class "SBF".}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS using Kirchoff-ray Mode approximation.
}
\details{
Calculates the theoretical TS using the Kirchoff-ray Mode model.
}
\references{
Clay C.S. and Horne J.K. (1994). Acoustic models of fish: The Atlantic cod
(Gadus morhua). Journal of the Acoustical Society of AMerica, 96, 1661-1668.
}
