% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{KRM}
\alias{KRM}
\title{Calculates the theoretical TS using Kirchoff-ray Mode approximation.}
\usage{
KRM(shape, c, rho, frequency, theta)
}
\arguments{
\item{shape}{Desired object/animal shape. Must be class "SBF".}

\item{c}{Sound speed of surrounding medium (m/s). Default value is 1500 m/s.}

\item{rho}{Density of medium (kg m^-1). Default value is 1030 kg m^-1.}

\item{frequency}{Frequency (Hz).}

\item{theta}{Orientation of the target relative to the transmit source 
(\eqn{\theta}). Broadside incidence is considered 90 degrees, or pi/2.
Default value is pi/2; input should be in radians.}
}
\value{
Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical TS using Kirchoff-ray Mode approximation.
}
\details{
Calculates the theoretical TS using the Kirchoff-ray Mode model.
}
\references{
Clay C.S. and Horne J.K. (1994). Acoustic models of fish: The Atlantic cod 
(Gadus morhua). Journal of the Acoustical Society of AMerica, 96, 1661-1668.
}
