% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_initialization.R
\name{krm_initialize}
\alias{krm_initialize}
\title{Initialize SBF-class object for KRM calculations.}
\usage{
krm_initialize(
  object,
  frequency,
  sound_speed_sw = 1500,
  density_sw = 1026,
  density_body = NULL,
  density_swimbladder = NULL,
  sound_speed_body = NULL,
  sound_speed_swimbladder = NULL,
  theta_body = NULL,
  theta_swimbladder = NULL
)
}
\arguments{
\item{object}{SBF-class object}

\item{frequency}{Frequency (Hz).}

\item{sound_speed_sw}{Seawater sound speed.}

\item{density_sw}{Seawater density.}

\item{density_body}{Optional flesh density input.}

\item{density_swimbladder}{Optional gas density input.}

\item{sound_speed_body}{Optional flesh sound speed input.}

\item{sound_speed_swimbladder}{Optional gas sound speed input.}

\item{theta_body}{Optional orientation input (relative to incident sound
wave).}

\item{theta_swimbladder}{Optional orientation input (relative to incident
sound wave).}
}
\description{
Initialize SBF-class object for KRM calculations.
}
