% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{linear}
\alias{linear}
\alias{dB}
\title{Convert backscatter values from log- to linear-domain}
\usage{
linear(value, coefficient = 10)

dB(value, coefficient = 10)
}
\arguments{
\item{value}{Logarithmic (e.g. TS or \eqn{S_V}) or linear (\eqn{\sigma_bs}) value}

\item{coefficient}{Optional. Numeric coefficient preceding the logarithm. Default is 10.}
}
\value{
Calculates the acoustic wavenumber based on the sound speed of water
}
\description{
Convert backscatter values from log- to linear-domain
}
\examples{
\dontrun{
TS <- -50 #dB re. 1 m^2
sigma_bs <- linear(TS) #convert to sigma_bs, 10^(TS/10)
print(sigma_bs)
TS_new <- dB(sigma_bs) #convert back to TS
print(TS_new)
}
}
