% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{linear}
\alias{linear}
\alias{dB}
\title{Convert backscatter values from log- to linear-domain}
\usage{
linear(value, coefficient = 10)

dB(value, coefficient = 10)
}
\arguments{
\item{value}{Logarithmic (e.g. TS) or linear (\eqn{\sigma_bs}) value}

\item{coefficient}{Optional. Numeric coefficient preceding the logarithm. Default is 10.}
}
\value{
Calculates the acoustic wavenumber based on the sound speed of water
}
\description{
Convert backscatter values from log- to linear-domain
}
