% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{OblateSpheroid}
\alias{OblateSpheroid}
\alias{OblateSpheroid-class}
\title{Oblate spheroidal body shape}
\description{
Oblate spheroidal body shape
}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes}
