% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{Pn}
\alias{Pn}
\title{Legendre Polynomial function (Pn) of the first kind.}
\usage{
Pn(n, x)
}
\arguments{
\item{n}{Polynomial degree.}

\item{x}{Interval bounded by -1 and 1}
}
\description{
Legendre Polynomial function (Pn) of the first kind.
}
