% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{Pn}
\alias{Pn}
\title{Legendre Polynomial function (Pn) of the first kind.}
\usage{
Pn(n, x)
}
\arguments{
\item{n}{Degree of Legendre polynomial}

\item{x}{Real value}
}
\value{
Returns a matrix array calculated by computing the outer product that generates
a series of values when evaluated over a vector.
}
\description{
Legendre Polynomial function (Pn) of the first kind.
}
