% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{pois}
\alias{pois}
\title{Calculate the Poisson's ratio (\eqn{\nu})}
\usage{
pois(K = NULL, E = NULL, G = NULL)
}
\arguments{
\item{K}{Bulk modulus (K, Pa).}

\item{E}{Young's modulus (E, Pa).}

\item{G}{Shear modulus (Pa).}
}
\value{
Poisson's ratio (\eqn{\nu}), dimensionless.
}
\description{
Calculates Poisson's ratio (\eqn{\nu}) from two of the three other elastic
moduli: bulk modulus (K), Young's modulus (E), or shear modulus (G).
Assumes 3D material properties.

The relationships used are:
\deqn{\nu = \frac{E}{2G} - 1}
\deqn{\nu = \frac{3K - 2G}{2(3K + G)}}
\deqn{\nu = \frac{3K - E}{6K}}
}
\keyword{elastic}
