% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{polynomial_cylinder}
\alias{polynomial_cylinder}
\title{Support function to generate polynomial deformed cylinder.}
\usage{
polynomial_cylinder(
  L,
  LAratio,
  N,
  g = 1.0357,
  h = 1.0279,
  theta = pi/2,
  pc = 3,
  curve = F,
  poly = c(0.83, 0.36, -2.1, -1.2, 0.63, 0.82, 0.64),
  object = "FLS"
)
}
\arguments{
\item{L}{Length of object (m).}

\item{LAratio}{Length:radius ratio.}

\item{N}{Number of cylinders.}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Orientation.}

\item{pc}{Radius of curvature ratio.}

\item{curve}{Curvature [T/F].}

\item{poly}{Vector of polynomial coefficients.}

\item{object}{Object class.}
}
\value{
Creates a polynomial deformed cylinder.
}
\description{
Support function to generate polynomial deformed cylinder.
}
