% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{PolynomialCylinder-class}
\alias{PolynomialCylinder-class}
\alias{PolynomialCylinder}
\title{Cylindrical body shape deformed using a polynomial}
\description{
Cylindrical body shape deformed using a polynomial
}
\seealso{
\code{\link{Shape}}
}
\keyword{internal}
\keyword{shapes,}
