% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{prolate_spheroid}
\alias{prolate_spheroid}
\title{Support function to generate prolate spheroid.}
\usage{
prolate_spheroid(
  L,
  LAratio,
  N,
  g = 1.0357,
  h = 1.0279,
  theta = pi/2,
  pc = 3,
  curve = F,
  object = "FLS"
)
}
\arguments{
\item{L}{Length of object (m).}

\item{LAratio}{Length:radius ratio.}

\item{N}{Number of cylinders.}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Orientation.}

\item{pc}{Radius of curvature ratio.}

\item{curve}{Curvature [T/F].}

\item{object}{Object class.}
}
\value{
Creates a prolate spheroid.
}
\description{
Support function to generate prolate spheroid.
}
