% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{prolate_spheroid}
\alias{prolate_spheroid}
\title{Creates a prolate spheroid.}
\usage{
prolate_spheroid(length, radius, length_radius_ratio, n_segments)
}
\arguments{
\item{length}{Semi-major axis length (m).}

\item{radius}{Semi-minor axis length (m). This can also be stylized as the
"maximum radius" of the scattering object.}

\item{length_radius_ratio}{Optional ratio input when radius is not explicitly
known.}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
1e2 segments.}
}
\value{
Creates the position vector for a prolate spheroid object of defined
semi-major and -minor axes.
}
\description{
Creates a prolate spheroid.
}
