% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{prolate_spheroid}
\alias{prolate_spheroid}
\title{Creates a prolate spheroid.}
\usage{
prolate_spheroid(
  length_body,
  radius_body,
  length_radius_ratio = NULL,
  n_segments = 18,
  length_units = "m",
  theta_units = "radians"
)
}
\arguments{
\item{length_body}{Semi-major axis length (m).}

\item{radius_body}{Semi-minor axis length (m). This can also be stylized as the
"maximum radius" of the scattering object.}

\item{length_radius_ratio}{Optional ratio input when radius is not explicitly
known.}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
18 segments.}

\item{length_units}{Units for body matrix (defaults to m).}

\item{theta_units}{Units for body orientation (defaults to radians).}
}
\value{
Creates the position vector for a prolate spheroid object of defined
semi-major and -minor axes.
}
\description{
Creates a prolate spheroid.
}
\seealso{
\code{\link{ProlateSpheroid-class}}
}
\keyword{shape_generation}
