% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R, R/model-psms.R
\name{PSMS}
\alias{PSMS}
\title{Calculates the theoretical target strength of a prolate spheroid using a
modal series solution}
\usage{
PSMS(object)

PSMS(object)
}
\arguments{
\item{object}{Prolate spheroid object.}
}
\value{
Target strength (TS, dB re: 1 m^2)

Target strength (TS, dB re: 1 m^2)
}
\description{
Calculates the theoretical target strength of a prolate spheroid using a
modal series solution

Calculates the theoretical target strength of a prolate spheroid using a
modal series solution
}
\references{
Furusawa, M. (1988). Prolate spheroidal models for predicting general trends
of fish target strength. Journal of the Acoustical Society of Japan, 9:
13-24.

Furusawa, M. (1988). Prolate spheroidal models for predicting general trends
of fish target strength. Journal of the Acoustical Society of Japan, 9:
13-24.
}
