% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{radians}
\alias{radians}
\title{Convert angular measurements from degrees to radians.}
\usage{
radians( x )
}
\arguments{
\item{x}{A real value in degrees}
}
\value{
Angle in radians.
}
\description{
Convert angular measurements from degrees to radians.
}
\examples{
orientation <- 90 # degrees
radians( orientation ) # this should return a value equal to pi / 2 radians#' 
}
