% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{reflection_coefficient}
\alias{reflection_coefficient}
\title{Plane wave/plane interface reflection coefficient}
\usage{
reflection_coefficient(interface1, interface2, mode = "DWBA")
}
\arguments{
\item{interface1}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (1)}

\item{interface2}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (2)}

\item{mode}{Two options: coefficient calculation for "DWBA" and "KRM"}
}
\description{
Plane wave/plane interface reflection coefficient
}
