% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge,FLS-method}
\alias{reforge,FLS-method}
\title{Reforge FLS-class object.}
\usage{
\S4method{reforge}{FLS}(object, new_length)
}
\arguments{
\item{object}{FLS-class object.}

\item{full_size}{New body length resize.}
}
\description{
Reforge FLS-class object.
}
