% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge,FLS-method}
\alias{reforge,FLS-method}
\title{Reforge FLS-class object.}
\usage{
\S4method{reforge}{FLS}(object, length, radius, length_radius_ratio_constant = T, n_segments)
}
\arguments{
\item{object}{FLS-class object.}

\item{length}{New body length resize.}

\item{radius}{New radius size}

\item{length_radius_ratio_constant}{Keep length-to-radius ratio based on new length}

\item{n_segments}{New number of segments}
}
\description{
Reforge FLS-class object.
}
