% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge,SBF-method}
\alias{reforge,SBF-method}
\title{Resizing function for swimbladdered targets}
\usage{
\S4method{reforge}{SBF}(
  object,
  body_scale = NULL,
  body_target = NULL,
  swimbladder_scale = NULL,
  swimbladder_target = NULL,
  isometric_body = TRUE,
  isometric_swimbladder = TRUE,
  maintain_ratio = TRUE,
  swimbladder_inflation_factor = 1,
  n_segments_body = NULL,
  n_segments_swimbladder = NULL
)
}
\arguments{
\item{object}{SBF-class object.}

\item{body_scale}{Proportional scaling to the body length, width, and height
dimensions. When a single value is supplied, all dimensions are scaled using
the same scaling factor. Otherwise, this input must be a named numeric
vector.}

\item{body_target}{Target dimensions (m) for the body length, width, and
height dimensions. This input must be a named numeric vector.}

\item{swimbladder_scale}{Proportional scaling to the swimbladder length,
width, and height dimensions. When a single value is supplied, all
dimensions are scaled using the same scaling factor. Otherwise, this input
must be a named numeric vector.}

\item{swimbladder_target}{Target dimensions (m) for the swimbladder length,
width, and height dimensions. This input must be a named numeric vector.}

\item{isometric_body}{Logical; maintain isometric scaling for body.}

\item{isometric_swimbladder}{Logical; maintain isometric scaling for bladder.}

\item{maintain_ratio}{Maintain size ratio between body and
swimbladder.}

\item{swimbladder_inflation_factor}{Proportional swimbladder volume where
the swimbladder x-axis origin and terminus are both held constant.}

\item{n_segments_body}{Number of segments along the body.}

\item{n_segments_swimbladder}{Number of segments along the bladder.}
}
\description{
Resizing function for swimbladdered targets
}
