% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge,SBF-method}
\alias{reforge,SBF-method}
\title{Resizing function for swimbladdered targets}
\usage{
\S4method{reforge}{SBF}(
  object,
  length_body = NA,
  width_body = NA,
  height_body = NA,
  length_bladder = NA,
  width_bladder = NA,
  height_bladder = NA,
  radius_body = NA,
  bladder_inflation_factor = 1,
  isometric_body = TRUE,
  isometric_bladder = TRUE,
  n_segments_body = NA,
  n_segments_bladder = NA
)
}
\arguments{
\item{object}{SBF-class object.}

\item{length_body}{Updated body length when applicable.}

\item{width_body}{Updated body width when applicable.}

\item{height_body}{Updated body height/depth when applicable.}

\item{length_bladder}{Updated bladder length when applicable.}

\item{width_bladder}{Updated bladder width when applicable.}

\item{height_bladder}{Updated bladder height/depth when applicable.}

\item{radius_body}{Updated body radius when applicable.}

\item{bladder_inflation_factor}{Proportional bladder volume.}

\item{isometric_body}{Logical; maintain isometric scaling for body.}

\item{isometric_bladder}{Logical; maintain isometric scaling for bladder.}

\item{n_segments_body}{Number of segments along the body.}

\item{n_segments_bladder}{Number of segments along the bladder.}
}
\description{
Resizing function for swimbladdered targets
}
