% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge,SBF-method}
\alias{reforge,SBF-method}
\title{Resizing function for swimbladdered targets}
\usage{
\S4method{reforge}{SBF}(
  object,
  length_body = NA,
  width_body = NA,
  height_body = NA,
  length_bladder = NA,
  width_bladder = NA,
  height_bladder = NA,
  radius_body = NA,
  bladder_inflation_factor = 1,
  isometric_body = T,
  isometric_bladder = T,
  n_segments_body = NA,
  n_segments_bladder = NA
)
}
\arguments{
\item{object}{SBF-class object.}

\item{full_size}{new factored resize of body.}

\item{bladder_size}{new factored resize of bladder.}
}
\description{
Resizing function for swimbladdered targets
}
