% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reforge,SBF-method}
\alias{reforge,SBF-method}
\title{Resizing function for swimbladdered targets}
\usage{
\S4method{reforge}{SBF}(object, full_size = 1, bladder_size = 1)
}
\arguments{
\item{object}{SBF-class object.}

\item{full_size}{new factored resize of body.}

\item{bladder_size}{new factored resize of bladder.}
}
\description{
Resizing function for swimbladdered targets
}
