% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reforge}
\alias{reforge}
\title{Resizing function for targets.}
\usage{
reforge(object, ...)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{...}{Additional inputs.}
}
\description{
Resizing function for targets.
}
