% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_reforge.R
\name{reforge}
\alias{reforge}
\title{Resize or reparameterize a scatterer object}
\usage{
reforge(object, ...)
}
\arguments{
\item{object}{A scatterer object.}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
Generic function to resize or reparameterize a scatterer object.
}
