% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{reset_model_registry}
\alias{reset_model_registry}
\title{Clear user-defined model registrations}
\usage{
reset_model_registry(remove_persisted = FALSE)
}
\arguments{
\item{remove_persisted}{Logical scalar. When \code{TRUE}, persisted user model
registrations are also deleted from the user's config file.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Clear user-defined model registrations
}
