% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{rho}
\alias{rho}
\title{Calculate the mass density material properties of a scatterer's tissue or
shell (kappa)}
\usage{
rho(interface1, interface2)
}
\arguments{
\item{interface1}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (1)}

\item{interface2}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (2)}
}
\description{
Calculate the mass density material properties of a scatterer's tissue or
shell (kappa)
}
