% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterer.R
\name{sbf_generate}
\alias{sbf_generate}
\title{Manually generate a SBF-class object.}
\usage{
sbf_generate(
  x_body,
  w_body,
  zU_body,
  zL_body,
  x_bladder,
  w_bladder,
  zU_bladder,
  zL_bladder,
  c_body,
  c_bladder,
  rho_body,
  rho_bladder,
  theta_body = pi/2,
  theta_bladder = pi/2,
  theta_units = "radians",
  length_units = "m",
  ID = NULL
)
}
\arguments{
\item{x_body}{Vector containing along-body axis (m).}

\item{w_body}{Vector containing across-body axis (m).}

\item{zU_body}{Vector containing dorsal-body axis (m).}

\item{zL_body}{Vector containing ventral-body axis (m).}

\item{x_bladder}{Vector containing along-bladder axis (m).}

\item{w_bladder}{Vector containing across-bladder axis (m).}

\item{zU_bladder}{Vector containing dorsal-bladder axis (m).}

\item{zL_bladder}{Vector containing ventral-bladder axis (m).}

\item{c_body}{Flesh sound speed (\ifelse{html}{\out{c;<sub>body</sub>}}{\eqn{c_{body}}}, m \ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}}).}

\item{c_bladder}{Bladder sound speed (c, m \eqn{s^-1}.}

\item{rho_body}{Flesh density (\ifelse{html}{\out{&rho;<sub>body</sub>}}{\eqn{\rho_{body}}}, kg \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}).}

\item{rho_bladder}{Bladder density (\eqn{\rho}, kg \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}).}

\item{theta_body}{Angle of body relative to wavefront (\eqn{\theta_body}, radians).}

\item{theta_bladder}{Angle of body relative to wavefront (\eqn{\theta_bladder}, radians).}

\item{theta_units}{Angular units.}

\item{length_units}{Angular units.}

\item{ID}{Angular units.}
}
\value{
Generates a SBF-class object.
}
\description{
Manually generate a SBF-class object.
}
