% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer_plot.R
\name{sbf_plot}
\alias{sbf_plot}
\title{Plotting for SBF-class objects}
\usage{
sbf_plot(
  object,
  type = "shape",
  nudge_y = 1.05,
  nudge_x = 1.01,
  x_units = "frequency"
)
}
\arguments{
\item{object}{SBF-class object.}

\item{type}{Toggle between body shape ("shape") or modeling results ("model")}

\item{nudge_y}{y-axis nudge.}

\item{nudge_x}{x-axis nudge.}

\item{x_units}{If "model" is selected, then toggle between frequency
("frequency", kHz) or ka ("ka").}

\item{aspect_ratio}{Aspect ratio setting ( defaults to "manual" for nudge_y
and nudge_x to apply; otherwise, input "auto").}

\item{y_units}{y-axis data selection (e.g. TS, sigma_bs -- defaults to TS)}
}
\description{
Plotting for SBF-class objects
}
