% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{sbf_show}
\alias{sbf_show}
\title{show(...) for SBF-class objects.}
\usage{
sbf_show(object)
}
\arguments{
\item{object}{SBF_class object.}
}
\value{
Called for its side effect of printing a formatted summary;
invisibly returns \code{NULL}.
}
\description{
show(...) for SBF-class objects.
}
\keyword{internal}
