% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\docType{class}
\name{Scatterer-class}
\alias{Scatterer-class}
\alias{Scatterer}
\title{Scatterer-class object for target strength estimation}
\description{
The \eqn{acousticTS} package uses a variety of defined S4-class objects
comprising different types of scatterers, such as fish with gas-filled
swimbladders (\link{SBF}) and fluid-like crustaceans
(\link{FLS}).
}
\section{Slots}{

\describe{
\item{\code{metadata}}{List containing relevant metadata}

\item{\code{model_parameters}}{Model parameters necessary for predicting TS}
}}

\section{Data Organization}{

\describe{
\item{\code{metadata}:}{A \code{list} comprising any identifying
information associated with the scatterer. The default metadata entry
includes \code{ID} that uses a default \code{character} value of
\code{"UID"} (i.e. \code{ID = "UID"}). This can otherwise be formatted
in any manner for book keeping purposes.}
\item{\code{body/bladder}:}{A \code{list} that includes information
relevant to the scatterer's position vector, material properties,
tilt/orientation, etc. For some scatterers, this may only include
\code{body}, but other targets may have an additional parameter such as
\code{bladder}. Generally, each entry includes:
\itemize{
\item{rpos}: the relevant position vector
(\ifelse{html}{\out{r<sub>0</sub>}}{\eqn{r_{0}}})
that includes axes such as \code{x}, \code{y}, \code{z}, etc., that depend
on the type of scatterer being used.
\item{\code{radius}:} in some cases this includes the radius measurements
for each cylinders depending on the type of scatterer object.
\item{\code{theta}:} the orientation of the scatterer relative to the
transmitting transducer or sound source
(\ifelse{html}{\out{&theta;<sub>animal</sub>}}{\eqn{\theta_{animal}}})
that can be represented either by degrees or radians, although all
functions require radians.
\item{\code{g, h}:} material properties that represent the density and
sound speed contrasts (g and h, respectively) relative to the
ambient/surrounding fluid. Some targets may instead have standard
sound speed
(\ifelse{html}{\out{c<sub>animal</sub>}}{\eqn{c_{animal}}},
m \ifelse{html}{\out{s<sup>-1</sup>}}{\eqn{s^{-1}}})
and density
(\ifelse{html}{\out{&rho;<sub>animal</sub>}}{\eqn{\rho_{animal}}},
kg \ifelse{html}{\out{m<sup>3</sup>}}{\eqn{m^3}}).
}
}
\item{\code{shape_parameters}:}{A \code{list} that includes metadata
pertaining to the shape of the scatterer and any other features of
interest (e.g. gas-filled swimbladder). Generally, each entry includes:
overall body length, the number of discrete cylinders that make up the
shape (if applicable), and the units related to both
\ifelse{html}{\out{&theta;<sub>animal</sub>}}{\eqn{\theta_{animal}}}
(e.g. rad, \ifelse{html}{\out{&deg;}}{\eqn{^\circ}}) and length
(e.g. mm, m).}
\item{\code{model_parameters}:}{A \code{list} that contains relevant model
parameterization once an object has been initialized for modeling
\ifelse{html}{\out{&sigma;<sub>bs</sub>}}{\eqn{\sigma_{bs}}}. This is
typically broken up into three categories:
\itemize{
\item{\code{parameters}:} A \code{list} that includes information such as
frequency (Hz), acoustic wavenumber (i.e. k), etc.
\item{\code{medium}:} A \code{data.frame} including information such as
the material properties of the ambient medium.
\item{\code{scatterer}:} A \code{list} containing summarized information
used to parameterize certain scattering models.}}
\item{\code{model}:}{A \code{list} that collects model results from one
or more models in the linear domain (i.e.
\ifelse{html}{\out{&sigma;<sub>bs</sub>}}{\eqn{\sigma_{bs}}}).
}
}
}

\section{Supported Scatterers}{

\itemize{
\item \code{Calibration spheres} (\link{CAL})
\item \code{Elastic-shelled scatterers} (\link{ESS})
\item \code{Fluid-like scatterers} (\link{FLS})
\item \code{Gas-filled scatterers} (\link{GAS})
\item \code{Swimbladdered fish} (\link{SBF})
}
}

\keyword{scatterer_type}
