% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SDWBA}
\alias{SDWBA}
\title{Calculates the theoretical TS of a fluid-like scatterer at a given frequency
using the stochastic distorted Born wave approximation (DWBA) model.}
\usage{
SDWBA(object)
}
\arguments{
\item{object}{FLS-class scatterer.}
}
\description{
Calculates the theoretical TS of a fluid-like scatterer at a given frequency
using the stochastic distorted Born wave approximation (DWBA) model.
}
\references{
Demer, D.A., and Conti, S.G. 2003. Reconciling theoretical versus empirical
target strengths of krill: effects of phase variability on the distorted-wave
Born approximation. ICES J. Mar. Sci., 60: 429-434.
}
