% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{class}
\name{Shape-class}
\alias{Shape-class}
\alias{Shape}
\title{Generic scattering shape object used throughout this package.}
\description{
A S4 class that provides slots to contain relevant shape data and metadata for
a variety of arbitrary and canonical shapes and geometries. See
\link{Scatterer} for a more detailed description on how
this S4 object interacts with generic \link{Scatterer}
objects.
}
\section{Slots}{

\describe{
\item{\code{position_matrix}}{Position matrix that provides the 2D representation of
the body shape}

\item{\code{shape_parameters}}{A list of additional shape specifications}
}}

\keyword{internal}
\keyword{shapes,}
