% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Shapely}
\alias{Shapely}
\title{Manipulate object}
\usage{

}
\arguments{
\item{shape}{Desired object/animal shape.}

\item{curve}{Curve (boolean, T/F).}

\item{pc}{Radius of curvature (pc)}

\item{theta}{Orientation}

\item{length}{New Length (m)}
}
\value{
Returns manipulated object.
}
\description{
Manipulate object
}
