% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{shear}
\alias{shear}
\title{Calculate the shear modulus (G)}
\usage{
shear(K = NULL, E = NULL, nu = NULL)
}
\arguments{
\item{K}{Bulk modulus (Pa).}

\item{E}{Young's modulus (Pa).}

\item{nu}{Poisson's ratio (Dimensionless).}
}
\value{
Shear modulus (G, Pa).
}
\description{
#' Calculates the shear modulus (G) from two of the three other elastic
moduli: bulk modulus (K), Young's modulus (E), or Poisson's ratio
(\eqn{\nu}). Assumes 3D material properties.

The relationships used are:
\deqn{G = \frac{3KE}{9K - E}}
\deqn{G = \frac{3K(1 - 2\nu)}{2(1 + \nu)}}
\deqn{G = \frac{E}{2(1 + \nu)}}
}
\keyword{elastic}
