% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{shear}
\alias{shear}
\title{Calculate the shear modulus (G).}
\usage{
shear(K, E, nu)
}
\arguments{
\item{K}{Bulk modulus (GPa).}

\item{E}{Young's modulus (GPa).}

\item{nu}{Poisson's ratio (unitless).}
}
\value{
Returns an estimate for the shear modulus (G).
}
\description{
Calculate the shear modulus (G) from two of the three other elastic
moduli to calculate the Lamé's parameter. When more than two values are input,
the function will default to using the bulk (K) and Young's (E) moduli. This
assumes that the input values represent 3D material properties.
}
