% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{shear}
\alias{shear}
\title{Calculate the shear modulus (G)}
\usage{
shear(K = NULL, E = NULL, nu = NULL)
}
\arguments{
\item{K}{Bulk modulus (Pa).}

\item{E}{Young's modulus (Pa).}

\item{nu}{Poisson's ratio (Dimensionless).}
}
\value{
Returns an estimate for the shear modulus (G).
}
\description{
Calculate the shear modulus (G) from two of the three other elastic
moduli to calculate the Lamé's parameter. When more than two
values are input, the function will default to using the bulk (K) and
Young's (E) moduli. This assumes that the input values represent 3D material
properties.
}
\keyword{elastic}
