% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,scatterer-method}
\alias{show,scatterer-method}
\title{Generic function for show(...) for different scatterers.}
\usage{
\S4method{show}{scatterer}(object)
}
\arguments{
\item{object}{Scattering object.}
}
\description{
Generic function for show(...) for different scatterers.
}
