% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_ts}
\alias{simulate_ts}
\title{Simulate target strength (TS) with flexible parameterization and batching}
\usage{
simulate_ts(
  object,
  frequency,
  model,
  n_realizations,
  parameters,
  batch_by = NULL,
  parallel = TRUE,
  n_cores = parallel::detectCores() - 1,
  verbose = TRUE
)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{frequency}{Frequency (Hz).}

\item{model}{Model name. If multiple models are specified, the output will
be a list of data frames, one for each model.}

\item{n_realizations}{Number of realizations and output TS values.}

\item{parameters}{List containing the values, distributions, or generating
functions of parameter values that inform the TS model.}

\item{batch_by}{Optional. Specifies which parameters in \code{parameters} to
batch over. Simulations will be run for all combinations of these parameter
values. Default is \code{NULL}.}

\item{parallel}{Logical; whether to parallelize the simulations. Default is
\code{TRUE}.}

\item{n_cores}{Optional. Number of CPU cores to use for parallelization.
Default is \code{parallel::detectCores() - 1}.}

\item{verbose}{Logical; whether to print progress and status messages to the
console. Default is \code{TRUE}.}
}
\value{
A data frame (or list of data frames) with simulation results.
}
\description{
Simulate target strength (TS) with flexible parameterization and batching
}
\details{
For example, if \code{batch_by = "length"} and \code{parameters["length"]}
is a vector of values, simulations will be run for each value of length
\code{n_realizations} times. If multiple parameters are specified in
\code{batch_by}, batching will occur over all combinations of their values.
}
\section{Parallelization}{

This function uses \code{pbapply::pblapply()} for parallelized simulation
with progress bars. On Windows, parallelization uses PSOCK clusters, which
require all necessary objects and packages to be exported to worker
processes. On Unix-like systems, forking is used, which is generally simpler.
}

\section{Performance Issues}{

Including too many parameters from \code{parameters}
within \code{batch_by} may cause significant performance issues or cause
\code{R} to crash. If intensive simulations are required, consider breaking
them into more manageable chunks
}

\section{DEPRECATION WARNING}{

The \code{simulate_ts} function will be deprecated in future versions and will be
replaced by the \code{anneal} function in future versions.
}

