% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{sphere}
\alias{sphere}
\title{Support function to generate sphere.}
\usage{
sphere(
  L,
  N,
  g = 1.0357,
  h = 1.0279,
  theta = pi/2,
  pc = 0,
  curve = F,
  object = "FLS"
)
}
\arguments{
\item{L}{Length of object (m).}

\item{N}{Number of cylinders.}

\item{g}{Density contrast.}

\item{h}{Sound speed contrast.}

\item{theta}{Orientation.}

\item{pc}{Radius of curvature ratio.}

\item{curve}{Curvature [T/F].}

\item{object}{Object class.}

\item{LAratio}{Length:radius ratio.}
}
\value{
Creates a sphere.
}
\description{
Support function to generate sphere.
}
