% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{sphere}
\alias{sphere}
\title{Creates a sphere.}
\usage{
sphere(radius, n_segments)
}
\arguments{
\item{radius}{Object radius (m).}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
1e3 segments.}
}
\value{
Creates position vector for a spherical object of a defined radius.
}
\description{
Creates a sphere.
}
