% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shape.R
\name{sphere}
\alias{sphere}
\title{Creates a sphere.}
\usage{
sphere(radius_body, n_segments, diameter_units )
}
\arguments{
\item{radius_body}{Object radius (m).}

\item{n_segments}{Number of segments to discretize object shape. Defaults to
1e2 segments.}

\item{diameter_units}{Default is "m" for meters}
}
\value{
Creates position vector for a spherical object of a defined radius.
}
\description{
Creates a sphere.
}
\seealso{
\code{\link{Sphere-class}}
}
\keyword{shape_generation}
