% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{target_strength}
\alias{target_strength}
\title{Wrapper function to model acoustic target strength}
\usage{
target_strength(object, frequency, model, ...)

target_strength(object, frequency, model, ...)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{frequency}{Frequency (Hz).}

\item{model}{Model name.}

\item{...}{Additional optional model inputs/parameters.}
}
\description{
Wrapper function to model acoustic target strength

Wrapper function to model acoustic target strength when multiple models are
defined.
}
