% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{target_strength}
\alias{target_strength}
\title{Model backscatter from a given target.}
\usage{
target_strength(object, frequency, model, ...)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{frequency}{Frequency (Hz).}

\item{model}{Model name.}

\item{...}{Additional optional model inputs/parameters.}
}
\description{
Model backscatter from a given target.
}
