% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{target_strength}
\alias{target_strength}
\title{Wrapper function to model acoustic target strength}
\usage{
target_strength(object, frequency, model, verbose = FALSE, ...)
}
\arguments{
\item{object}{Scatterer-class object.}

\item{frequency}{Frequency (Hz).}

\item{model}{Model name.}

\item{verbose}{Prints current procedural step occurring from model initialization
to calculating TS. Defaults to FALSE.}

\item{...}{Additional optional model inputs/parameters.}
}
\description{
Wrapper function to model acoustic target strength
}
