% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\name{transmission_coefficient}
\alias{transmission_coefficient}
\title{Transmission coefficient for transmission between two mediums}
\usage{
transmission_coefficient(interface1, interface2)
}
\arguments{
\item{interface1}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (1)}

\item{interface2}{Dataframe object containing density (kg/m^3) and sound
speed (m/s) values for a boundary/interface (2)}
}
\description{
Transmission coefficient for transmission between two mediums
}
