% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TS2sigma}
\alias{TS2sigma}
\title{Toggle between log- and linear-domain for backscatter values.}
\usage{
To convert from TS to \eqn{\sigma_bs}
TS2sigma(x)
To convert from \eqn{\sigma_bs} to TS
sigma2TS(x)
}
\arguments{
\item{x}{A real value in degrees or radians}
}
\value{
Converts TS to sigma_bs and vice versa.
}
\description{
Toggle between log- and linear-domain for backscatter values.
}
\examples{
TS <- -85 #dB re: 1 m^2
(sigma <- TS2sigma(TS)) #m^2; convert TS to sigma_bs
[1] 3.162278e-09
sigma2TS(sigma) #dB re: 1 m^2; convert sigma_bs to TS
[1] -85
}
