% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDWBA.R, R/utilities.R
\name{vecnorm}
\alias{vecnorm}
\title{Calculates the Euclidean norm of a vector.}
\usage{
vecnorm(x)

vecnorm(x)
}
\arguments{
\item{x}{A vector with numeric, real values.}

\item{x}{A vector with numeric, real values.}
}
\value{
Calculates the Euclidean norm of a vector.

Calculates the Euclidean norm of a vector.
}
\description{
Calculates the Euclidean norm of a vector.

Calculates the Euclidean norm of a vector.
}
\examples{
values <- c(1,2,3)
vecnorm(values)
[1] 3.741657
values <- c(1,2,3)
vecnorm(values)
[1] 3.741657
}
