% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{vecnorm}
\alias{vecnorm}
\title{Calculates the Euclidean norm of a matrix input.}
\usage{
vecnorm(x)
}
\arguments{
\item{x}{A matrix with numeric, real values.}
}
\value{
Calculates the Euclidean norm of a vector.
}
\description{
Calculates the Euclidean norm of a matrix input.
}
\examples{
values <- matrix(c(1,2,3), ncol=3)
vecnorm(values)
# 3.741657
}
