% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_sphere.R
\name{ya}
\alias{ya}
\title{Calculate Bessel function of the second kind.}
\usage{
ya(l,n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Bessel function of the second kind (Y_a). Functions are referenced from Amost D.E., "AMOS, A Portable Package 
for Bessel Functions of a Complex Argument and Nonnegative Order", http://netlib.org/amos
}
\description{
Calculate Bessel function of the second kind.
}
\examples{
l <- 1.5
n <- 1
ya(l,n)
[1] -1.102496
}
