% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{yc}
\alias{yc}
\alias{ycd}
\title{Cylindrical Bessel (Neumann) function of the second kind and its derivative}
\usage{
yc(l, n)

ycd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the cylindrical Bessel function of the first kind (Y_v) and its
respective derivatives.
}
\description{
Calculate the cylindrical Bessel function of the first kind (yc) and its
first (ycd) derivatives.
}
\references{
Amos D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
