% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{yc}
\alias{yc}
\alias{ycd}
\title{Wrapper function for the Cylindrical Bessel (Neumann) function of the second kind (Yc)}
\usage{
yc( l , n )

ycd( l , n )
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the Bessel function of the second kind (Y_v). Functions are
referenced from Amost D.E., "AMOS, A Portable Package
for Bessel Functions of a Complex Argument and Nonnegative Order",
http://netlib.org/amos
}
\description{
Wrapper function for the Cylindrical Bessel (Neumann) function of the second kind (Yc)

First derivative of Cylindrical Bessel function of the second kind (Ycd)
}
