% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustics.R
\encoding{UTF-8}
\name{young}
\alias{young}
\title{Calculate Young's modulus (E).}
\usage{
young(K = NULL, G = NULL, nu = NULL)
}
\arguments{
\item{K}{Bulk modulus (Pa).}

\item{G}{Shear modulus (Pa).}

\item{nu}{Poisson's ratio (Dimensionless).}
}
\value{
Young's modulus (E, Pa).
}
\description{
Calculates Young's modulus (E) from two of the three other elastic moduli:
bulk modulus (K), shear modulus (G), or Poisson's ratio (\eqn{\nu}).
Assumes 3D material properties.

The relationships used are:
\deqn{E = \frac{9KG}{3K + G}}
\deqn{E = 3K(1 - 2\nu)}
\deqn{E = 2G(1 + \nu)}
}
\keyword{elastic}
