% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_functions.R
\name{ys}
\alias{ys}
\alias{ysd}
\title{Calculate spherical Bessel function of the second kind.}
\usage{
ys(l,n)

ysd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
ys(l,n) calculates the spherical Bessel function of the second kind (zbesy
AMOS routine, Yv)
}
\description{
This function calculates the spherical Bessel functions of the second kind.
Referenced from Amost D.E.,
"AMOS, A Portable Package for Bessel Functions of a Complex Argument and
Nonnegative Order", http://netlib.org/amos
}
\examples{
l <- 1
n <- 1
ys(l,n)
# -1.381773
}
