% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel.R
\name{ys}
\alias{ys}
\alias{ysd}
\title{Spherical Bessel function of the second kind and its respective derivative}
\usage{
ys(l, n)

ysd(l, n)
}
\arguments{
\item{l}{An integer or fractional order}

\item{n}{A complex or real argument}
}
\value{
Calculates the spherical Bessel function of the first kind (js) and its
first (ysd) derivative.
}
\description{
Spherical Bessel function of the second kind and its respective derivative
}
\references{
Amos D.E., "AMOS, A Portable Package for Bessel Functions of a Complex
Argument and Nonnegative Order", http://netlib.org/amos
}
